% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brin_parca}
\alias{brin_parca}
\title{BRIN *PARCA* layer}
\format{
An `sf` object with 114 obs and 18 colums.
}
\usage{
brin_parca
}
\description{
An example of *PARCA* layer for the Brin forest.
}
\keyword{datasets}
