% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brin_ua}
\alias{brin_ua}
\title{BRIN *UA* layer}
\format{
An `sf` object with 215 obs and 45 colums.
}
\usage{
brin_ua
}
\description{
An example of *UA* layer for the Brin forest.
The information contained therein is entirely imaginary.
It does not reflect the description or management of the forest at all.
}
\keyword{datasets}
