% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_names}
\alias{clean_names}
\title{Clean data frame names}
\usage{
clean_names(x)
}
\arguments{
\item{x}{\code{character} String to clean}
}
\value{
A cleaned string.
}
\description{
Internal helper used to clean df names
}
\keyword{internal}
