% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-gis.R
\name{create_empty_sf}
\alias{create_empty_sf}
\title{Create an empty sf object}
\usage{
create_empty_sf(geom_type, ...)
}
\arguments{
\item{geom_type}{`character`; one of the geometry types supported by `sf`: `POLYGON`,
`LINESTRING`, `POINT`, `MULTIPOLYGON`, `MULTILINESTRING`, `MULTIPOINT`.}

\item{...}{Named attributes with their classes. For example, you can pass
arguments like `PLACETTE = character(0)`. Each argument must be named according
to the attribute, with the corresponding empty vector of the appropriate class.}
}
\value{
An `sf` object
}
\description{
This function creates an empty `sf` object with the desired geometry type
(point, line, polygon, etc.).
}
\examples{
\dontrun{
empty_sf <- create_empty_sf(
  "LINESTRING",
  PLACETTE = character(0),
  TSE_VOL = numeric(0)
)
}

}
