% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legal_entity.R
\name{download_legal_entity}
\alias{download_legal_entity}
\title{Download and cache cadastral data for legal entities ("Personnes Morales")}
\usage{
download_legal_entity(cache = seq_cache("legal_entity")$path, verbose = TRUE)
}
\arguments{
\item{cache}{\code{character}; Storage directory. Defaults to the user cache
directory.}

\item{verbose}{\code{logical}; If \code{TRUE}, display progress and informational messages.}
}
\value{
\code{character};
}
\description{
This function downloads the official "Personnes Morales" datasets published
by the French Ministry of Economy and Finance. These datasets contain
cadastral and property information (parcels and buildings) owned by
legal entities. The data are automatically cached locally to avoid
repeated downloads.
}
\examples{
\dontrun{
download_legal_entity(c(29, "08"))
}

}
