% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drias.R
\name{drias_ombro}
\alias{drias_ombro}
\title{Compute ombrothermic climatology summaries}
\usage{
drias_ombro(txt)
}
\arguments{
\item{txt}{\code{character} string. Path to a DRIAS \code{.txt} file.}
}
\value{
A \code{data.frame} containing monthly averages of temperature and precipitation
for each period.
}
\description{
Computes monthly ombrothermic summaries from DRIAS extraction
}
\examples{
\dontrun{
ombro <- drias_ombro(clim)
}

}
