% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drias.R
\name{drias_read_table}
\alias{drias_read_table}
\title{Parse DRIAS raw data from a text file}
\usage{
drias_read_table(txt)
}
\arguments{
\item{txt}{\code{character} string. Path to a DRIAS \code{.txt} file.}
}
\value{
A \code{data.frame} containing raw data.
}
\description{
Extracts raw data from a DRIAS text file.
}
