% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geol-bdcharm50.R
\name{find_bdcharm50_zip}
\alias{find_bdcharm50_zip}
\title{Find BDCharm50 archive for a department}
\usage{
find_bdcharm50_zip(dep, urls)
}
\arguments{
\item{dep}{Department code.}

\item{urls}{Named `character` vector of archive URLs, usually returned by
[get_bdcharm50_url()].}
}
\value{
A `character(1)` ZIP filename.
}
\description{
Finds the unique BDCharm50 ZIP archive matching a department code.
}
\keyword{internal}
