% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legal_entity.R
\name{format_legal_entity}
\alias{format_legal_entity}
\title{Format and aggregate legal-entity data}
\usage{
format_legal_entity(legal_entity, code_insee = NULL, verbose = FALSE)
}
\description{
Format and aggregate legal-entity data
}
\keyword{internal}
