% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessibility.R
\name{get_accessibility}
\alias{get_accessibility}
\title{Retrieve forest accessibility features (porter or skidder)}
\usage{
get_accessibility(
  x,
  type = c("porteur", "skidder"),
  buffer = 1000,
  verbose = TRUE
)
}
\arguments{
\item{x}{An \code{sf} object defining the input area of interest.}

\item{type}{\code{character} Accessibility type. One of \code{"porter"} or \code{"skidder"}.}

\item{buffer}{\code{numeric}; Buffer around \code{x} (in \strong{meters}) used to enlarge
the download area.}

\item{verbose}{\code{logical}; If \code{TRUE}, display progress and informational
messages.}
}
\value{
An \code{sf} object containing accessibility features, or \code{NULL} if none found.
}
\description{
Builds a convex buffer around the input geometry and retrieves accessibility
features for the requested machine type.
}
