% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ars.R
\name{get_ars}
\alias{get_ars}
\title{Download ARS vector layer from IGN WFS}
\usage{
get_ars(x, key = c("captage", "ppi", "ppr", "ppe"), buffer = 500)
}
\arguments{
\item{x}{\code{sf} or \code{sfc}; Geometry located in France.}

\item{key}{\code{character}; key identifying layer to download. Must be one of
\code{"captage"}, \code{"ppe"}, \code{"ppi"} or \code{"ppr"}.}

\item{buffer}{\code{numeric}; Buffer around \code{x} (in \strong{meters}) used to enlarge
the download area.}
}
\description{
ARS layers are typically used to assess sanitary constraints, ensure
compliance with water-protection regulations, and understand environmental
risks near captages in a given study area.
}
\details{
Four ARS layers are currently supported:
\itemize{
\item \strong{\code{captage}}: Drinking-water abstraction points declared by partner
organisations (collectivités, services d’eau). These represent the physical
points where water is captured for human consumption.
\item \strong{\code{ppi}} : The most restrictive protection zone, directly surrounding the
captage, usually fenced and strictly controlled("Périmètre de Protection Immédiate")
\item \strong{\code{ppr}} : An intermediate protection zone between the PPI and PPE,
with moderate regulatory constraints ("Périmètre de Protection Rapprochée")
\item \strong{\code{ppe}}: A broad protection zone around a captage where activities may
be regulated to safeguard water quality ("Périmètre de Protection Élargi")
}
}
