% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geol-bdcharm50.R
\name{get_bdcharm50_url}
\alias{get_bdcharm50_url}
\title{List BDCharm50 archive URLs}
\usage{
get_bdcharm50_url(dep)
}
\arguments{
\item{dep}{French num department}
}
\value{
A named \code{character} vector. Names are ZIP filenames and values are
download URLs.
}
\description{
Scrapes the BDCharm50 download directory and returns available ZIP archive
URLs, named by archive filename.
}
\keyword{internal}
