% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cog.R
\name{get_cog}
\alias{get_cog}
\title{Get COG administrative datasets}
\usage{
get_cog(cache = seq_cache("cog")$path, update = FALSE, verbose = TRUE)
}
\arguments{
\item{cache}{Cache directory. If \code{NULL}, uses package cache directory.}

\item{update}{Force re-download even if cached.}

\item{verbose}{\code{logical}; If \code{TRUE}, display progress and informational messages.}
}
\value{
Named list of data frames:
\itemize{
\item \code{com}: communes
\item \code{dep}: departments
\item \code{reg}: regions
}
}
\description{
Download and cache French COG datasets (communes, departments, regions)
from data.gouv.fr. Cached data are reused unless \code{update = TRUE}.
}
\examples{
\dontrun{
cog <- get_cog()
cog <- get_cog(update = TRUE)
}

}
