% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curves.R
\name{get_curves}
\alias{get_curves}
\title{Retrieve contour lines around an area}
\usage{
get_curves(x, buffer = 5000, verbose = TRUE)
}
\arguments{
\item{x}{\code{sf} or \code{sfc}; Input area.}

\item{buffer}{\code{numeric}; Buffer around \code{x}, in meters.}

\item{verbose}{\code{logical}; If \code{TRUE}, display messages.}
}
\value{
An \code{sf} line layer, or \code{NULL} if no contour line is found.
}
\description{
Retrieves contour lines within a buffered envelope around an area.
}
