% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alti-rgealti.R
\name{get_dsm}
\alias{get_dsm}
\title{Download Digital Surface Model (DSM) raster from IGN RGE ALTI}
\usage{
get_dsm(x, buffer = 200, res = 1, crs = 2154, verbose = TRUE)
}
\arguments{
\item{x}{\code{sf} or \code{sfc}; Geometry located in France.}

\item{buffer}{\code{numeric}; Buffer around \code{x} (in \strong{meters}) used to enlarge
the download area.}

\item{res}{\code{numeric}; resolution specified in the units of the coordinate
system (see \code{\link[happign:get_wms_raster]{happign::get_wms_raster()}})}

\item{crs}{\code{numeric} or \code{character}; CRS of the returned raster (see
\code{\link[happign:get_wms_raster]{happign::get_wms_raster()}})}

\item{verbose}{\code{logical}; If \code{TRUE}, display progress and informational messages.}
}
\value{
\code{SpatRaster} object from \code{terra} package
}
\description{
Compatibility wrapper around \code{\link[=get_rge]{get_rge()}} for an MNS.
}
\seealso{
\code{\link[happign:get_wms_raster]{happign::get_wms_raster()}}
}
