% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro.R
\name{get_hydro_poly}
\alias{get_hydro_poly}
\title{Retrieve hydrographic polygons around an area}
\usage{
get_hydro_poly(x, buffer = 1000)
}
\arguments{
\item{x}{An `sf` object used as the input area.}

\item{buffer}{`numeric`; Buffer around `x` (in **meters**) used to enlarge}
}
\value{
An `sf` object containing hydrographic polygons with two fields:
  * `TYPE` - hydrographic class
    - `RSO` = Reservoir or water tower
    - `SFP` = Permanent hydrographic surface
    - `SFI` = Intermittent hydrographic surface
  * `NATURE` - Original BDTOPO nature field
  * `NAME` - Official hydrographic name (when available)
}
\description{
Retrieve hydrographic polygons around an area
}
\details{
The function retrieves BDTOPO layers within a convex buffer
around `x`, assigns the hydrographic types, and combines them into
a single layer.
}
