% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legal_entity.R
\name{get_legal_entity}
\alias{get_legal_entity}
\title{Create a forest matrice for legal entity from insee code}
\usage{
get_legal_entity(x, cache = seq_cache("legal_entity")$path, verbose = TRUE)
}
\arguments{
\item{x}{\code{character}; Code(s) INSEE or code(s) department (see
[Rsequoia2::get_cog())}

\item{cache}{\code{character}; Storage directory. Defaults to the user cache
directory.}

\item{verbose}{\code{logical}; If \code{TRUE}, display progress and informational messages.}
}
\description{
Generates a forest matrice used to store general forest
information (e.g., \code{IDENTIFIANT}, \code{PROPRIETAIRE}) and cadastral attributes
(\code{CODE_INSEE}, \code{PREFIXE}, \code{SECTION}, \code{NUMERO}, \code{LIEU_DIT}) for legal
entity.
}
\examples{
\dontrun{
pm <- get_legal_entity(c("29158", "08185"))
}

}
