% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parca.R
\name{get_lieux_dits}
\alias{get_lieux_dits}
\title{Retrieve a "Lieud-dit" from Etalab}
\usage{
get_lieux_dits(idu)
}
\arguments{
\item{idu}{\code{character} Cadastral parcel identifier.}
}
\value{
An \code{sf} object containing the parcel geometry.
}
\description{
Retrieve a "Lieud-dit" from Etalab
}
