% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old.R
\name{get_old}
\alias{get_old}
\title{Retrieve OLD features around an area}
\usage{
get_old(x, buffer = 1000, verbose = TRUE)
}
\arguments{
\item{x}{An \code{sf} object defining the input area of interest.}

\item{buffer}{\code{numeric}; Buffer around \code{x} (in \strong{meters}) used to enlarge}

\item{verbose}{\code{logical}; If \code{TRUE}, display progress and informational messages.}
}
\value{
An \code{sf} object containing OLD features.
}
\description{
Builds a convex buffer around the input geometry, retrieves OLD
features and returns an \code{sf} point layer.
}
\details{
The function creates a convex buffer around the input geometry \code{x}
and retrieves OLD features before returns as a single \code{sf} point layer.
}
