% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ortho.R
\name{get_ortho}
\alias{get_ortho}
\title{Download orthophotos from the IGN WMTS}
\usage{
get_ortho(
  x,
  type = c("irc", "rgb"),
  buffer = 200,
  zoom = 12,
  crs = 2154,
  overwrite = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{x}{\code{sf} or \code{sfc}; Geometry located in France.}

\item{type}{\code{character}; One of \code{"irc"} or \code{"rgb"}.}

\item{buffer}{\code{numeric}; Buffer around \code{x}, in meters.}

\item{zoom}{\code{integer}; WMTS zoom level.}

\item{crs}{CRS of the returned raster.}

\item{overwrite}{\code{logical}; If \code{TRUE}, overwrite existing files.}

\item{verbose}{\code{logical}; If \code{TRUE}, display progress and informational
messages.}
}
\value{
A \code{SpatRaster}.
}
\description{
Download orthophotos from the IGN WMTS
}
