% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parca.R
\name{get_parca_bdp}
\alias{get_parca_bdp}
\title{Retrieve a BDP parcel geometry from BDP}
\usage{
get_parca_bdp(idu)
}
\arguments{
\item{idu}{\code{character} Cadastral parcel identifier.}
}
\value{
An \code{sf} object containing the parcel geometry.
}
\description{
Retrieve a BDP parcel geometry from BDP
}
