% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patrimony.R
\name{get_patrimony}
\alias{get_patrimony}
\title{Download patrimony vector layer}
\usage{
get_patrimony(x, key, buffer = 1000, verbose = TRUE)
}
\arguments{
\item{x}{\code{sf} or \code{sfc}; Geometry located in France.}

\item{key}{\code{character}; Layer to download.
Must be one of from \code{get_keys("pat")}}

\item{buffer}{\code{numeric}; Buffer around \code{x} (in \strong{meters}) used to enlarge
the download area.}

\item{verbose}{\code{logical}; If \code{TRUE}, display progress and informational messages.}
}
\value{
\code{sf} object from \code{sf} package
}
\description{
Downloads a vector layer with \code{frheritage} for the area covering \code{x}
expanded with a buffer.
}
