% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedology.R
\name{get_pedology}
\alias{get_pedology}
\title{Retrieve pedology polygon features around an area}
\usage{
get_pedology(x)
}
\arguments{
\item{x}{An `sf` object defining the input area of interest.}
}
\value{
An `sf` object containing pedology polygon features
intersecting the input geometry, with additional surface fields.
}
\description{
Retrieves pedological polygon features from the INRA soil map
intersecting an area of interest and computes surface attributes.
}
\details{
The function retrieves pedology polygon features from the
`INRA.CARTE.SOLS:geoportail_vf` layer intersecting the input geometry `x`.
The resulting geometries are intersected with `x`, cast to polygons,
and surface attributes are computed using `ua_generate_area()`.
}
\seealso{
[ua_generate_area()]
}
