% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prsf.R
\name{get_prsf}
\alias{get_prsf}
\title{Retrieve PRSF point features around an area}
\usage{
get_prsf(x, buffer = 5000, verbose = TRUE)
}
\arguments{
\item{x}{An \code{sf} or \code{sfc} object defining the input area.}

\item{buffer}{\code{numeric}; Buffer around \code{x}, in meters.}

\item{verbose}{\code{logical}; If \code{TRUE}, display messages.}
}
\value{
An \code{sf} point layer, or \code{NULL} if no PRSF feature is found.
}
\description{
Retrieves PRSF point features around an area of interest.
}
