% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alti-rgealti.R
\name{get_rge}
\alias{get_rge}
\title{Download an altimetry raster from IGN RGE ALTI}
\usage{
get_rge(
  x,
  key = c("mnt", "mns", "mnh"),
  buffer = 200,
  res = 1,
  crs = 2154,
  minmax = c(0, 50),
  verbose = TRUE
)
}
\arguments{
\item{x}{\code{sf} or \code{sfc}; Geometry located in France.}

\item{key}{\code{character}; RGE ALTI product to retrieve. One of \code{"mnt"},
\code{"mns"} or \code{"mnh"}.}

\item{buffer}{\code{numeric}; Buffer around \code{x} (in \strong{meters}) used to enlarge
the download area.}

\item{res}{\code{numeric}; resolution specified in the units of the coordinate
system (see \code{\link[happign:get_wms_raster]{happign::get_wms_raster()}})}

\item{crs}{\code{numeric} or \code{character}; CRS of the returned raster (see
\code{\link[happign:get_wms_raster]{happign::get_wms_raster()}})}

\item{minmax}{\code{numeric}; Accepted MNH range.}

\item{verbose}{\code{logical}; If \code{TRUE}, display progress and informational messages.}
}
\value{
\code{SpatRaster} object from \code{terra} package
}
\description{
Downloads an MNT, MNS or MNH from the IGN WMS service for the area covering
\code{x}. The MNH is computed from the downloaded MNT and MNS.
}
\seealso{
\code{\link[happign:get_wms_raster]{happign::get_wms_raster()}}
}
