% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vege.R
\name{get_vege_line}
\alias{get_vege_line}
\title{Retrieve forest vegetation lines around an area}
\usage{
get_vege_line(x, buffer = 1000, clip = TRUE, tol = 500, poly = NULL)
}
\arguments{
\item{x}{\code{sf} or \code{sfc}; Geometry located in France.}

\item{buffer}{\code{numeric}; Buffer around \code{x}, in meters.}

\item{clip}{\code{logical}; If \code{TRUE}, remove the forest area itself.}

\item{tol}{\code{numeric}; Minimum area threshold in square meters.}

\item{poly}{Optional preloaded vegetation polygon layer.}
}
\value{
An \code{sf} line layer.
}
\description{
Retrieve forest vegetation lines around an area
}
