% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vege.R
\name{get_vege_point}
\alias{get_vege_point}
\title{Generate vegetation point features around an area}
\usage{
get_vege_point(x, buffer = 1000, clip = TRUE)
}
\arguments{
\item{x}{\code{sf} or \code{sfc}; Geometry located in France.}

\item{buffer}{\code{numeric}; Buffer around \code{x}, in meters.}

\item{clip}{\code{logical}; If \code{TRUE}, remove the forest area itself.}
}
\value{
An \code{sf} point layer.
}
\description{
Generate vegetation point features around an area
}
