% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vege.R
\name{get_vege_poly}
\alias{get_vege_poly}
\title{Retrieve forest vegetation polygons around an area}
\usage{
get_vege_poly(x, buffer = 1000, clip = TRUE, tol = 500)
}
\arguments{
\item{x}{\code{sf} or \code{sfc}; Geometry located in France.}

\item{buffer}{\code{numeric}; Buffer around \code{x}, in meters.}

\item{clip}{\code{logical}; If \code{TRUE}, remove the forest area itself.}

\item{tol}{\code{numeric}; Minimum area threshold in square meters.}
}
\value{
An \code{sf} polygon layer. An empty standardized layer is returned
when no vegetation is found.
}
\description{
Retrieve forest vegetation polygons around an area
}
