% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/road.R
\name{get_vides}
\alias{get_vides}
\title{Retrieve cadastral gaps around the \emph{PARCA}}
\usage{
get_vides(x)
}
\arguments{
\item{x}{An \code{sf} object used as the input area.
It must contain a source field with \code{bdp} or \code{etalab} value.}
}
\value{
An \code{sf} object of type \code{POLYGON} containing cadastral gaps
with standardized fields, including:
\itemize{
\item \code{TYPE} - Empty, to complete
\item \code{NATURE} - Empty
\item \code{NAME} - Empty, to complete
\item \code{SOURCE} - Empty
}
}
\description{
Retrieve cadastral gaps around the \emph{PARCA}
}
