% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voids.R
\name{get_voids}
\alias{get_voids}
\title{Retrieve cadastral gaps around the _PARCA_}
\usage{
get_voids(x)
}
\arguments{
\item{x}{An `sf` object used as the input area.
It must contain a source field with `bdp` or `etalab` value.}
}
\value{
An `sf` object of type `POLYGON` containing cadastral gaps
  with standardized fields, including:
  * `TYPE` - Empty, to complete
  * `NATURE` - Empty
  * `NAME` - Empty, to complete
  * `SOURCE` - Empty
}
\description{
Retrieve cadastral gaps around the _PARCA_
}
