% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{idu_build}
\alias{idu_build}
\title{Build IDU}
\usage{
idu_build(dep, com, prefix, section, numero)
}
\arguments{
\item{dep}{\code{character} dep code}

\item{com}{\code{character} com code}

\item{prefix}{\code{character} prefic code}

\item{section}{\code{character} section code}

\item{numero}{\code{character} numero code}
}
\description{
Internal helper used to build idu
}
\keyword{internal}
