% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteo-france.R
\name{mf_get_climate_fiche}
\alias{mf_get_climate_fiche}
\title{Download climatological station fiche}
\usage{
mf_get_climate_fiche(x, dirname, verbose = TRUE)
}
\arguments{
\item{x}{\code{sf} or \code{sfc}}

\item{dirname}{\code{character}; Output directory for the generated workbook.}

\item{verbose}{\code{logical}; If \code{TRUE}, display progress and informational messages.}
}
\value{
Path to the downloaded PDF
}
\description{
Downloads the climatological fiche (PDF) of the nearest Meteo-France station
to the provided geometry.
}
\details{
The fiche contains station metadata and long-term climatological summaries.
}
\examples{
\dontrun{
mf_get_climate_fiche(parca, dirname = "METEO")
}

}
