% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteo-france.R
\name{mf_get_climatology}
\alias{mf_get_climatology}
\title{Download monthly climatology from Meteo-France}
\usage{
mf_get_climatology(x, cache = seq_cache("meteo")$path, verbose = TRUE)
}
\arguments{
\item{x}{\code{sf} or \code{sfc}.}

\item{cache}{\code{character}; Storage directory. Defaults to the user cache
directory (see \code{\link[tools:R_user_dir]{tools::R_user_dir()}})}

\item{verbose}{\code{logical}; If \code{TRUE}, display progress and informational messages.}
}
\value{
A \code{data.frame} containing monthly climatology records for the nearest stations.
}
\description{
Downloads monthly climatological data for the \emph{3 nearest meteorological stations}.
}
\details{
Cached files are reused to avoid redundant downloads.
}
\examples{
\dontrun{
clim <- mf_get_climatology(parca)
}

}
