% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteo-france.R
\name{mf_precipitation}
\alias{mf_precipitation}
\title{Compute annual precipitation summaries}
\usage{
mf_precipitation(clim)
}
\arguments{
\item{clim}{\code{data.frame} Raw climatology data from \code{\link[=mf_get_climatology]{mf_get_climatology()}}.}
}
\value{
A \code{data.frame} containing annual precipitation averages.
}
\description{
Computes annual precipitation averages from raw climatology data.
}
\details{
Precipitation values are computed from monthly records and aggregated by year.
}
\examples{
\dontrun{
precip <- mf_precipitation(clim)
}

}
