% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legal_entity.R
\name{normalize_legal_entity}
\alias{normalize_legal_entity}
\title{Normalize legal-entity matrice}
\usage{
normalize_legal_entity(legal_entity, verbose = FALSE)
}
\description{
Normalize legal-entity matrice
}
\keyword{internal}
