% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm-get.R
\name{normalize_pm}
\alias{normalize_pm}
\title{Normalize legal entity parcels}
\usage{
normalize_pm(pm, verbose = FALSE)
}
\arguments{
\item{pm}{A legal entity parcel data frame returned by \code{\link[=get_pm]{get_pm()}}.}

\item{verbose}{\code{logical}; If \code{TRUE}, display progress and informational messages.}
}
\value{
A list with:
\describe{
\item{m}{Unique normalized parcel table.}
\item{m_detail}{Detailed normalized parcel table with culture information.}
}
}
\description{
Convert legal entity parcel data to the Sequoia parcel format and build a
detailed parcel table.
}
