% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalize_txt}
\alias{normalize_txt}
\title{Normalize text}
\usage{
normalize_txt(x)
}
\arguments{
\item{x}{\code{character} String to clean}
}
\value{
A cleaned string.
}
\description{
Internal helper used to normalize string for query
}
\keyword{internal}
