% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pad_left}
\alias{pad_left}
\title{Left-pad a string with zeros}
\usage{
pad_left(x, width, fill = "0")
}
\arguments{
\item{x}{`character` String to pad}

\item{width}{`integer` Desired total width}

\item{fill}{`character` Filler to pad with}
}
\value{
A zero-padded character string.
}
\description{
Internal helper used to format fixed-width codes.
}
\keyword{internal}
