% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ua.R
\name{parca_to_ua}
\alias{parca_to_ua}
\title{Create _UA_ sf object from _PARCA_ sf object}
\usage{
parca_to_ua(parca)
}
\arguments{
\item{parca}{`sf` Object from [Rsequoia2::seq_parca()] containing cadastral
parcels.}
}
\value{
An `sf` object containing the analysis units.
}
\description{
Create _UA_ sf object from _PARCA_ sf object
}
