% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-gis.R
\name{poly_to_line}
\alias{poly_to_line}
\title{Get a border layer from polygons}
\usage{
poly_to_line(x)
}
\arguments{
\item{x}{\verb{sf POLYGONS}, using a projected CRS}
}
\value{
An \code{sf} object with two \code{MULTILINESTRING} geometry is returned
}
\description{
This function convert polygon to line and remove shared lines
}
\note{
If the polygon layer contains topology errors (such as contiguous
polygons not sharing exactly the same boundary)
}
