% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parca.R
\name{read_etalab}
\alias{read_etalab}
\title{Read Etalab cadastral data}
\usage{
read_etalab(insee, layer = c("parcelles", "lieux_dits"))
}
\arguments{
\item{insee}{\code{character}. INSEE commune code(s).}

\item{layer}{\code{character}. Etalab layer to read. Either \code{"parcelles"} or
\code{"lieux_dits"}.}
}
\value{
An \code{sf} object containing the requested Etalab layer.
}
\description{
Downloads and combines Etalab cadastral GeoJSON data for one or more
communes.
}
\keyword{internal}
