% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrice.R
\name{read_matrice}
\alias{read_matrice}
\title{Read matrice and format for Rsequoia2}
\usage{
read_matrice(dirname = ".")
}
\arguments{
\item{dirname}{\code{character}; Path to the Sequoia project directory.
Defaults to the current working directory.}
}
\value{
\code{data.frame} formated as Rsequoia2 matrice
}
\description{
Read matrice and format for Rsequoia2
}
