% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parca.R
\name{remove_holes}
\alias{remove_holes}
\title{Remove Holes in a \emph{PARCA} Layer}
\usage{
remove_holes(parca)
}
\arguments{
\item{parca}{An \code{sf} object, typically produced by \code{\link[=seq_parca]{seq_parca()}},
containing cadastral parcels.}
}
\value{
An updated \code{sf} object based on \code{parca}, with holes replaced
by geometries without attributes.
}
\description{
This function fills holes in a \emph{PARCA} layer by adding geometries
corresponding to the missing areas.
}
