% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roadway.R
\name{road_difference}
\alias{road_difference}
\title{Compute hierarchical road differences}
\usage{
road_difference(buffers, order)
}
\arguments{
\item{buffers}{A named list of \code{sf} polygon buffers.}

\item{order}{Character vector. Defines the processing order of road types.}
}
\value{
An \code{sf} object with combined polygons after hierarchical differences.
}
\description{
Performs sequential differences between buffered road polygons according to a given hierarchy.
}
\keyword{internal}
