% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roadway.R
\name{road_mask}
\alias{road_mask}
\title{Create a road external mask}
\usage{
road_mask(x, dist)
}
\arguments{
\item{x}{An \code{sf} object with road lines.}

\item{dist}{Numeric. Buffer distance in layer units.}
}
\value{
An \code{sf} object containing a polygon mask of the road network.
}
\description{
Generates a mask polygon around road lines, including buffered intersection points.
}
\keyword{internal}
