% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legal_entity.R
\name{search_legal_entity}
\alias{search_legal_entity}
\title{Helpers to search within a forest matrix}
\usage{
search_legal_entity(x, prop = NULL, lieu_dit = NULL)
}
\arguments{
\item{x}{`data.frame`; The matrix containing parcel legal entity information
in a format readable by SEQUOIA.}

\item{prop}{`character`; Owner name(s) used to further filter results. Defaults to `NULL`.
Multiple owners can be specified.}

\item{lieu_dit}{`character`; Locality name(s) used to further filter results. Defaults to `NULL`.
Multiple locality names can be specified.}
}
\value{
A `data.frame`.
}
\description{
This function allows searching for parcels in a forest matrix based on
owner names or location (lieu-dit).
}
\details{
The search relies on a text-normalization step applied to both the inputs and
the corresponding columns of the matrice. This makes the search robust
to accents, punctuation, spacing irregularities and case differences.

**Examples of normalization:**
  - `"Etat / Forets"` -> `"ETATFORETS"`
  - `"  Le Bois-de l'Orme  "` -> `"LEBOISDELORME"`
  - `"Societe du Chene"` -> `"SOCIETEDUCHENE"`
}
\examples{
\dontrun{
legal_entity <- get_legal_entity(29158)
search_mat <- search_legal_entity(legal_entity, c("penmarch", "guenole"))
}

}
