% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legal_entity.R
\name{search_matrice}
\alias{search_matrice}
\title{Helpers to search within a forest matrix}
\usage{
search_matrice(matrice, prop = NULL, lieu_dit = NULL)
}
\arguments{
\item{matrice}{\code{data.frame}; The matrix containing parcel information in a
format readable by SEQUOIA.}

\item{prop}{\code{character}; Owner name(s) used to further filter results. Defaults to \code{NULL}.
Multiple owners can be specified.}

\item{lieu_dit}{\code{character}; Locality name(s) used to further filter results. Defaults to \code{NULL}.
Multiple locality names can be specified.}
}
\value{
A \code{data.frame}.
}
\description{
This function allows searching for parcels in a forest matrix based on
owner names or location (lieu-dit).
}
\examples{
\dontrun{
# Recherche des parcelles dans la matrice pour le département 01
pm29158 <- get_pm("29", "29158")

# Recherche des parcelles pour une commune spécifique et un propriétaire
cad <- search_pm(pm29158, prop = c("COMMUNE"))
cad <- search_pm(pm29158, lieu_dit = "ECKMUHL")
cad <- search_pm(pm29158, prop = c("COMMUNE"), lieu_dit = "ECKMUHL")
}

}
