% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm-get.R
\name{select_pm}
\alias{select_pm}
\title{Select legal entities interactively}
\usage{
select_pm(pm)
}
\arguments{
\item{pm}{A data frame returned by \code{\link[=search_pm]{search_pm()}}.}
}
\value{
A character vector of selected legal entity names.
}
\description{
Build a summary by legal entity and prompt the user to select one or more
legal entities.
}
