% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ars.R
\name{seq_ars}
\alias{seq_ars}
\title{Download ARS vector layers for a Sequoia project}
\usage{
seq_ars(
  dirname = ".",
  buffer = 1000,
  key = get_keys("mnhn"),
  verbose = TRUE,
  overwrite = FALSE
)
}
\arguments{
\item{dirname}{\code{character} Directory where the matrice file is located.
Defaults to the current working directory.}

\item{buffer}{\code{numeric}; Buffer around \code{x} (in \strong{meters}) used to enlarge
the download area.}

\item{key}{\code{character}; List of ARS layer identifiers to download. If not
provided, the function automatically select all ARS layers defined in the
Sequoia configuration (\code{inst/config/seq_layers.yaml})}

\item{verbose}{\code{logical} If \code{TRUE}, display messages.}

\item{overwrite}{\code{logical} If \code{TRUE}, file is overwritten.}
}
\value{
A named list of file paths written by \code{\link[=seq_write]{seq_write()}}, one per layer.
}
\description{
Downloads one or several vector layers from the IGN WFS service
for \code{ars} layer(s) of a Sequoia project.
}
\details{
This function is a convenience wrapper looping over ars layers (\code{"captage"},
\code{"ppi"}, \code{"ppr"}, \code{"ppe"}), allowing the user to download all products in
one call and automatically write them to the project directory using
\code{\link[=seq_write]{seq_write()}}.
}
\seealso{
\code{\link[=get_ars]{get_ars()}}, \code{\link[=seq_write]{seq_write()}}
}
