% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundaries.R
\name{seq_boundaries}
\alias{seq_boundaries}
\title{Create \emph{FOREST} and \emph{OWNER} object from \emph{PARCA} for a Sequoia project}
\usage{
seq_boundaries(dirname = ".", tol = 10, verbose = TRUE, overwrite = FALSE)
}
\arguments{
\item{dirname}{\code{character}; Path to the Sequoia project directory.
Defaults to the current working directory.}

\item{tol}{\code{numeric}; buffer tolerance in meter used to close small gaps.
Any gap narrower than \code{2 * tol} is filled, causing polygons on both sides to
be merged.}

\item{verbose}{\code{logical}; If \code{TRUE}, display progress and informational
messages.}

\item{overwrite}{\code{logical}; If \code{TRUE}, overwrite existing files.}
}
\value{
A named list
}
\description{
Aggregates parcel geometries to build forest and owner boundaries.
}
\details{
\strong{About \code{tol}}
The gap-closing logic is based on a French forestry rule stating that a forest
split by a path or narrow road is legally considered a single continuous stand

To reflect this, small linear gaps are removed using the sequence
\emph{buffer -> union -> unbuffer}, applied with the distance given by \code{tol}.
Any gap narrower than \code{2 * tol} is closed.

\strong{Output layers}
The function produces six layers:
\itemize{
\item \code{FOREST_poly}   : Forest boundaries with aggregated surface.
\strong{\code{tol} is applied} to remove small gaps.
\item \code{FOREST_line}   : Forest boundaries as linestrings.
\item \code{FOREST_point}  : Centroids of forest polygons.
\item \code{OWNER_poly}    : Boundaries aggregated per owner.
\strong{\code{tol} is \emph{not} applied} to avoid merging polygons across owners.
\item \code{OWNER_line}    : Owner boundaries as linestrings.
\item \code{OWNER_point}   : Centroids of owner polygons.
}
}
