% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brgm.R
\name{seq_geol}
\alias{seq_geol}
\title{Create a geology layer for a Sequoia project from BRGM data}
\usage{
seq_geol(dirname = ".", cache = NULL, verbose = FALSE, overwrite = FALSE)
}
\arguments{
\item{dirname}{\code{character} Directory where the matrice file is located.
Defaults to the current working directory.}

\item{cache}{\code{character}; Storage directory. Defaults to the user cache
directory (see \code{\link[tools:userdir]{tools::R_user_dir()}}).}

\item{verbose}{\code{logical} If \code{TRUE}, display messages.}

\item{overwrite}{\code{logical} If \code{TRUE}, overwrite zipfile.}
}
\value{
An invisible \code{sf} object containing the merged geology layer.
}
\description{
Uses the \emph{PARCA} from Sequoia to determine which French departments are
involved, downloads the corresponding BRGM geology datasets, and
builds a single geology layer for the project.
}
\details{
Created layer and its \emph{QML style file} are automatically written to the
project directory using \code{\link[=seq_write]{seq_write()}}.
}
\keyword{internal}
