% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate.R
\name{seq_climat}
\alias{seq_climat}
\title{Generate meteorological outputs}
\usage{
seq_climat(dirname = ".", cache = NULL, verbose = TRUE, overwrite = FALSE)
}
\arguments{
\item{dirname}{\code{character} Directory where the matrice file is located.
Defaults to the current working directory.}

\item{cache}{\code{character}; Storage directory. Defaults to the user cache
directory (see \code{\link[tools:userdir]{tools::R_user_dir()}})}

\item{verbose}{\code{logical} If \code{TRUE}, display messages.}

\item{overwrite}{\code{logical} If \code{TRUE}, file is overwritten.}
}
\value{
A named \code{list} of file paths :
\itemize{
\item `"fiche.meteo"``: Path to the downloaded climatological station PDF
\item `"meteofrance"``: Path to the generated Excel workbook
}
}
\description{
Downloads and formats meteorological data from Météo-France and DRIAS.
This includes download of the nearest climatological station fiche (PDF),
ombrothermic summaries, precipitation statistics. Results are written
to an Excel workbook based on the internal template \code{CLIMAT_MF.xlsx}.
}
