% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{seq_file_exist}
\alias{seq_file_exist}
\title{Check if a file already exist}
\usage{
seq_file_exist(key, dirname)
}
\arguments{
\item{key}{key}

\item{dirname}{dirname}
}
\description{
Internal helper used to check if a file already exist
}
\keyword{internal}
