% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{seq_files_rename}
\alias{seq_files_rename}
\title{Rename matching files}
\usage{
seq_files_rename(path, old_id, new_id)
}
\arguments{
\item{path}{Character. Directory path.}

\item{old_id}{Character. Identifier to replace.}

\item{new_id}{Character. Replacement identifier.}
}
\value{
Logical vector returned by \code{\link[=file.rename]{file.rename()}}.
}
\description{
Rename files recursively by replacing \code{old_id} with \code{new_id} in file names.
}
\keyword{internal}
