% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{seq_gpkg_id_rename}
\alias{seq_gpkg_id_rename}
\title{Rename GeoPackage identifier}
\usage{
seq_gpkg_id_rename(path, new_id)
}
\arguments{
\item{path}{Character. Path to a GeoPackage.}

\item{new_id}{Character. New identifier value.}
}
\value{
A list with \code{file}, \code{status}, and \code{message}.
}
\description{
Update the Sequoia identifier field in one GeoPackage.
}
\keyword{internal}
