% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infra.R
\name{seq_infra}
\alias{seq_infra}
\title{Generate infrastructure layers for a Sequoia project}
\usage{
seq_infra(dirname = ".", buffer = 1000, verbose = TRUE, overwrite = FALSE)
}
\arguments{
\item{dirname}{\code{character}; Path to the Sequoia project directory.
Defaults to the current working directory.}

\item{buffer}{\code{numeric}; Buffer around \code{x} (in \strong{meters}) used to enlarge}

\item{verbose}{\code{logical}; If \code{TRUE}, display progress and informational
messages.}

\item{overwrite}{\code{logical}; If \code{TRUE}, overwrite existing files.}
}
\value{
Invisibly returns a named list of written file paths.
}
\description{
Retrieves infrastructure polygon, line and point layers around the project
area and writes them to the Sequoia project directory.
}
\seealso{
\code{\link[=get_infra_poly]{get_infra_poly()}}, \code{\link[=get_infra_line]{get_infra_line()}}, \code{\link[=get_infra_point]{get_infra_point()}}
}
