% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteo-france.R
\name{seq_meteo_france}
\alias{seq_meteo_france}
\title{Download Meteo-France data for a Sequoia project}
\usage{
seq_meteo_france(
  dirname = ".",
  cache = NULL,
  verbose = TRUE,
  overwrite = FALSE
)
}
\arguments{
\item{dirname}{\code{character}; Path to the Sequoia project directory.
Defaults to the current working directory.}

\item{cache}{\code{character}; Storage directory. Defaults to the user cache
directory (see \code{\link[tools:R_user_dir]{tools::R_user_dir()}})}

\item{verbose}{\code{logical}; If \code{TRUE}, display progress and informational
messages.}

\item{overwrite}{\code{logical}; If \code{TRUE}, overwrite existing files.}
}
\value{
Invisibly returns a named list containing:
\itemize{
\item \code{fiche.meteo}: climatological station PDF;
\item \code{meteofrance}: generated Meteo-France workbook.
}
}
\description{
Downloads the nearest climatological station fiche and generates the
Meteo-France workbook for the project area.
}
