% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parca.R
\name{seq_parca}
\alias{seq_parca}
\title{Download, enrich and write cadastral geometries}
\usage{
seq_parca(dirname = ".", verbose = TRUE, overwrite = FALSE)
}
\arguments{
\item{dirname}{\code{character}; Path to the Sequoia project directory.
Defaults to the current working directory.}

\item{verbose}{\code{logical}; If \code{TRUE}, display progress and informational
messages.}

\item{overwrite}{\code{logical}; If \code{TRUE}, overwrite existing files.}
}
\value{
An \code{sf} object
}
\description{
This function reads the \code{matrice.xlsx} file from a project directory,
downloads the geometry of each parcel, enriches the dataset with information
from the matrice, and writes the resulting layer to disk.
}
\details{
The resulting object is returned invisibly as an \code{sf} polygons layer.
The output file is automatically written into the working directory defined
by \code{dirname}.

\strong{Automatic "lieu-dit" completion}
If the function detects rows in the matrice where the field \code{"LIEU_DIT"}
is missing, the corresponding "lieu-dit" values will be downloaded
automatically and added to the output.
Existing \code{"LIEU_DIT"} values are \strong{never overwritten}.
}
