% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ua-create.R
\name{seq_parca_to_ua}
\alias{seq_parca_to_ua}
\title{Create analysis units layer}
\usage{
seq_parca_to_ua(dirname = ".", verbose = TRUE, overwrite = FALSE)
}
\arguments{
\item{dirname}{\code{character}; Path to the Sequoia project directory.
Defaults to the current working directory.}

\item{verbose}{\code{logical}; If \code{TRUE}, display progress and informational
messages.}

\item{overwrite}{\code{logical}; If \code{TRUE}, overwrite existing files.}
}
\value{
An \code{sf} object
}
\description{
This function reads the cadastral layer file from a project directory
and create the analysis units layer.
}
\details{
The resulting object is returned invisibly as an \code{sf} polygons layer.
The output file is automatically written into the working directory defined
by \code{dirname}.
}
