% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{seq_summary}
\alias{seq_summary}
\title{Summarize information for a Sequoia folder}
\usage{
seq_summary(dirname = ".", verbose = TRUE)
}
\arguments{
\item{dirname}{\code{character}; Path to the Sequoia project directory.
Defaults to the current working directory.}

\item{verbose}{\code{logical}; If \code{TRUE}, display progress and informational
messages.}
}
\value{
\code{list} of \code{data.frame}
}
\description{
Generate synthetic table from Sequoia layers and write them as \code{.xlsx}
inside the current Sequoia dir. Useful for redaction of management document.
}
\details{
Available tables:
\itemize{
\item \code{"CAD"}: Cadastral parcels
\item \code{"CAD_COM"}: Surfaces by communes
\item \code{"PF"}: Surfaces by forest parcels
\item \code{"SSPF"}: Surfaces by forest sub-parcels
\item \code{"CAD_PLT"}: Link between cadastral parcels and forest parcels
\item \code{"OCCUPATION"}: Surfaces by land use
\item \code{"STATION"}: Surfaces by station
\item \code{"GEOL_BDCHARM50"}: Surfaces by geology (source: BDCHARM50)
\item \code{"GEOL_CARHAB"}: Surfaces by geology (source: CARHAB)
\item \code{"PEDO"}: Surfaces by pedology type
\item \code{"PLT_PF"}: Link between stand type and forest parcels
\item \code{"PF_PLT"}: Link between forest parcels and stand type
\item \code{"GESTION"}: Surfaces by management type
\item \code{"ALTI_PF"}: Altimetry recap by forest parcels (max, min, mean)
\item \code{"EXPO_PF"}: Exposition recap by forest parcels
\item \code{"PENTE_PF"}: Slope recap by forest parcels
}
}
