% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write.R
\name{seq_write}
\alias{seq_write}
\title{Write a spatial object based on a layer key}
\usage{
seq_write(x, key, dirname = ".", id = NULL, verbose = FALSE, overwrite = FALSE)
}
\arguments{
\item{x}{An \code{sf} object (for vector outputs) or a \code{SpatRaster} (for raster outputs).}

\item{key}{\code{character} Name of a layer key to match against the entries
defined in \code{inst/config/seq_layers.yaml}. (see \emph{Details} for partial matching).}

\item{dirname}{\code{character}; Path to the Sequoia project directory.
Defaults to the current working directory.}

\item{id}{\code{character}. Project identifier used to prefix the output filename.
If \code{NULL}, the identifier is automatically inferred. The function first
attempts to extract a single, non-missing value from the \code{IDENTIFIANT}
field of layer \code{x}; if this fails, it then looks for a valid identifier in
a \verb{*_matrice.xlsx} file found in \code{dirname}. If no unique identifier can be
determined, the filename is left unchanged.}

\item{verbose}{\code{logical}; If \code{TRUE}, display progress and informational
messages.}

\item{overwrite}{\code{logical}; If \code{TRUE}, overwrite existing files.}
}
\value{
Invisibly returns the filepath used for writing.
}
\description{
Write a spatial object based on a layer key
}
