% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ua.R
\name{ua_check_area}
\alias{ua_check_area}
\title{Update cadastral area values in \emph{UA} using \emph{PARCA}}
\usage{
ua_check_area(ua, parca, verbose = FALSE)
}
\arguments{
\item{ua}{\code{sf} Object from \code{\link[=parca_to_ua]{parca_to_ua()}} containing analysis
units}

\item{parca}{\code{sf} Object from \code{\link[=seq_parca]{seq_parca()}} containing cadastral
parcels.}

\item{verbose}{\code{logical} If \code{TRUE}, display progress messages.}
}
\value{
Updated \code{ua} object.
}
\description{
This function compares cadastral area values between \emph{UA} and \emph{PARCA},
and updates \emph{UA} wherever discrepancies are detected.
}
