% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ua-parca.R
\name{ua_check_idu}
\alias{ua_check_idu}
\title{Check cadastral IDU consistency between \emph{UA} and \emph{PARCA}}
\usage{
ua_check_idu(ua, parca, verbose = FALSE)
}
\arguments{
\item{ua}{\code{sf} Object containing analysis units.}

\item{parca}{\code{sf} Object containing cadastral parcels.}

\item{verbose}{\code{logical}; If \code{TRUE}, display progress and informational messages.}
}
\value{
\code{TRUE} if IDU values are consistent; \code{FALSE} otherwise,
with CLI warnings.
}
\description{
Checks that all \emph{PARCA} IDU values are present in \emph{UA}, and that all \emph{UA}
IDU values exist in \emph{PARCA}.
}
