% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ua-occupation.R
\name{ua_check_non_wooded_threshold}
\alias{ua_check_non_wooded_threshold}
\title{Check non-wooded surface threshold}
\usage{
ua_check_non_wooded_threshold(ua, threshold = 0.1, verbose = TRUE)
}
\arguments{
\item{ua}{\code{sf} object containing UA polygons.}

\item{threshold}{Maximum allowed non-wooded ratio. Defaults to \code{0.10}.}

\item{verbose}{\code{logical}; If \code{TRUE}, display progress and informational messages.}
}
\value{
\code{TRUE} if the threshold is respected, otherwise \code{FALSE}.
}
\description{
Checks that non-wooded submitted surfaces do not exceed 10\% of submitted
DGD surfaces.
}
