% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ua.R
\name{ua_get_areas}
\alias{ua_get_areas}
\title{Calculate areas in UA}
\usage{
ua_get_areas(ua, verbose = TRUE)
}
\arguments{
\item{ua}{\code{sf} object containing analysis units.}

\item{verbose}{\code{logical} If \code{TRUE}, display progress messages.}
}
\value{
\code{sf} object with updated areas
}
\description{
Computes the corrected cadastral areas for units of analysis according to
cartographic area.
}
