% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ua-occupation.R
\name{ua_to_occupation}
\alias{ua_to_occupation}
\title{Aggregate \emph{UA} surfaces by occupation status}
\usage{
ua_to_occupation(ua, verbose = TRUE)
}
\arguments{
\item{ua}{\code{sf} object containing analysis units.}

\item{verbose}{\code{logical}; If \code{TRUE}, display progress and informational messages.}
}
\value{
An \code{sf} object aggregated by occupation status.
}
\description{
Aggregates corrected surface areas from a \emph{UA} layer by grouping rows
according to DGD-submitted and wooded status.
}
\seealso{
\code{\link[=seq_occupation]{seq_occupation()}}, \code{\link[=seq_field]{seq_field()}}, \code{\link[=seq_normalize]{seq_normalize()}}
}
