% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parcels.R
\name{ua_to_pf}
\alias{ua_to_pf}
\title{Aggregate _UA_ surfaces at the `parcelle` field level}
\usage{
ua_to_pf(ua)
}
\arguments{
\item{ua}{`sf` object containing analysis units}
}
\value{
`sf` object where polygons are aggregated at the _PF_ level, with
corrected surfaces.
}
\description{
Aggregates corrected surface areas from a _UA_ layer by
grouping rows according to the parcel unit field, as defined by the
configuration returned by [seq_field()].
}
\details{
This function is typically used internally by [seq_parcels()], but may also
be called directly when a PF-level summary of a UA dataset is required.
}
\seealso{
[ua_to_sspf()], [seq_parcels()], [seq_field()], [seq_normalize()]
}
