% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ua-parcels.R
\name{ua_to_sspf}
\alias{ua_to_sspf}
\title{Aggregate _UA_ surfaces at the `sspf` field level}
\usage{
ua_to_sspf(ua)
}
\arguments{
\item{ua}{`sf` object containing analysis units}
}
\value{
An `sf` object where polygons and descriptive fields are aggregated
at the _SSPF_ level.
}
\description{
Aggregates corrected surface areas from a _UA_ layer by
grouping rows according to the `sspf` field, as
defined by the configuration returned by [seq_field()].
}
\details{
In addition to summing surfaces, descriptive fields are preserved.

This function is used internally by [seq_parcels()], but may be called
directly when a SSPF-level summary of UA data is needed.
}
\seealso{
[ua_to_pf()], [seq_parcels()], [seq_field()], [seq_normalize()]
}
