% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wooded.R
\name{ua_to_wooded}
\alias{ua_to_wooded}
\title{Aggregate _UA_ surfaces by wooded / non-wooded status}
\usage{
ua_to_wooded(ua)
}
\arguments{
\item{ua}{`sf` object containing analysis units.}
}
\value{
An `sf` object where polygons are aggregated by wooded status, with
corrected surfaces.
}
\description{
Aggregates corrected surface areas from a _UA_ layer by
grouping rows according to the wooded status field, as defined by the
configuration returned by [seq_field()].
}
\details{
This function is typically used internally by [seq_wooded()], but may also
be called directly when a summary of wooded versus non-wooded areas is
required.
}
\seealso{
[seq_wooded()], [seq_field()], [seq_normalize()]
}
