% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ua-parca.R
\name{ua_update_parca_fields}
\alias{ua_update_parca_fields}
\title{Update _UA_ fields from _PARCA_}
\usage{
ua_update_parca_fields(ua, parca)
}
\arguments{
\item{ua}{`sf` Object containing analysis units.}

\item{parca}{`sf` Object containing cadastral parcels.}
}
\value{
A `sf` _UA_ object with cadastral fields updated from _PARCA_.
}
\description{
Refreshes _UA_ cadastral fields from _PARCA_ using the IDU field as key.
}
\details{
This function assumes that _UA_ IDU values are already correct.
}
