% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ua.R
\name{ua_update_surfcad}
\alias{ua_update_surfcad}
\title{Update cadastral area values in UA using PARCA}
\usage{
ua_update_surfcad(ua, parca, verbose = TRUE)
}
\arguments{
\item{ua}{\code{sf} object containing analysis units;
must contain the idu and surf_cad fields.}

\item{parca}{\code{sf} Object containing cadastral parcels;
must contain the idu and surf_cad fields.}

\item{verbose}{\code{logical} If \code{TRUE}, display progress messages.}
}
\value{
Updated \code{ua} object.
}
\description{
This function compares cadastral area values between UA and PARCA,
and updates UA wherever discrepancies are detected.
}
