% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update_parca}
\alias{update_parca}
\title{Update `R_SEQUOIA` _PARCA_ layer to `Rsequoia2`}
\usage{
update_parca(parca, id)
}
\arguments{
\item{parca}{`sf` object containing cadastrals parcels}

\item{id}{`character` Identifier of the project}
}
\value{
An `sf` object containing cadastral parcels, normalized for
`Rsequoia2` use.
}
\description{
Update `R_SEQUOIA` _PARCA_ layer to `Rsequoia2`
}
