% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update_ua}
\alias{update_ua}
\title{Update `R_SEQUOIA` _UA_ layer to `Rsequoia2`}
\usage{
update_ua(ua, id)
}
\arguments{
\item{ua}{`sf` object containing analysis units}

\item{id}{`character` Identifier of the project}
}
\value{
An `sf` object containing analysis units, normalized for
`Rsequoia2` use.
}
\description{
Update `R_SEQUOIA` _UA_ layer to `Rsequoia2`
}
