% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roadway.R
\name{voids_to_roadway}
\alias{voids_to_roadway}
\title{Get polygons from the completed cadastral gaps layer}
\usage{
voids_to_roadway(x)
}
\arguments{
\item{x}{An \code{sf} polygon object from \code{\link[=get_voids]{get_voids()}}
completed by user}
}
\value{
An \code{sf} object with POLYGON or MULTIPOLYGON geometries normalized as
road polygons.
}
\description{
Cleans polygon topology and normalizes the result as road polygons.
}
\keyword{internal}
