% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrice.R
\name{create_matrice}
\alias{create_matrice}
\title{Create a  forest matrice}
\usage{
create_matrice(
  dirname = ".",
  id = "MY_FOREST",
  overwrite = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{dirname}{`character` Path to the directory. Defaults to the current
working directory.}

\item{id}{`character` Identifier of the forest. Typically the name of the
forest}

\item{overwrite}{`logical` If `TRUE`, filename is overwritten.}

\item{verbose}{`logical` If `TRUE`, display messages.}
}
\value{
Invisibly returns the path to the created file.
}
\description{
Generates a default forest matrice as excel file used to store general forest
information (e.g., `IDENTIFIANT`, `PROPRIETAIRE`) and cadastral attributes
(`CODE_INSEE`, `PREFIXE`, `SECTION`, `NUMERO`, `LIEU_DIT`).
}
\examples{
\dontrun{
create_matrice("~/matrice.xlsx", overwrite = TRUE)
}
}
