% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_matrice.R
\name{create_matrice}
\alias{create_matrice}
\title{Create a  forest matrice}
\usage{
create_matrice(
  dirname = ".",
  id = "MY_FOREST",
  overwrite = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{dirname}{\code{character} Path to the directory. Defaults to the current
working directory.}

\item{id}{\code{character} Identifier of the forest. Typically the name of the
forest}

\item{overwrite}{\code{logical} If \code{TRUE}, filename is overwritten.}

\item{verbose}{\code{logical} If \code{TRUE}, display messages.}
}
\value{
Invisibly returns the path to the created file.
}
\description{
Generates a default forest matrice as excel file used to store general forest
information (e.g., \code{IDENTIFIANT}, \code{PROPRIETAIRE}) and cadastral attributes
(\code{CODE_INSEE}, \code{PREFIXE}, \code{SECTION}, \code{NUMERO}, \code{LIEU_DIT}).
}
\examples{
\dontrun{
create_matrice("~/matrice.xlsx", overwrite = TRUE)
}
}
