% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brgm.R
\name{download_brgm}
\alias{download_brgm}
\title{Download BRGM "BD Charm 50" harmonised geology ZIP files}
\usage{
download_brgm(dep, cache = NULL, verbose = FALSE, overwrite = FALSE)
}
\arguments{
\item{dep}{\code{character} or \code{numeric}; Department code (see \link[happign:dep_2025]{happign::dep_2025}).}

\item{cache}{\code{character}; Storage directory. Defaults to the user cache
directory (see \code{\link[tools:userdir]{tools::R_user_dir()}}).}

\item{verbose}{\code{logical} If \code{TRUE}, display messages.}

\item{overwrite}{\code{logical} If \code{TRUE}, overwrite zipfile.}
}
\value{
Character vector of directories where ZIPs are stored.
}
\description{
Downloads BRGM ZIP archives for one french department. Files are stored
in a persistent cache directory (\code{R_user_dir("Rsequoia2", "cache")}).
}
\details{
Existing ZIP files are never re-downloaded.
}
