% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alti-lidar.R
\name{download_lidar}
\alias{download_lidar}
\title{Download LIDAR HD tiles}
\usage{
download_lidar(
  x,
  key = c("mnt", "mnh"),
  cache = NULL,
  overwrite = FALSE,
  verbose = TRUE,
  max_tries = 10
)
}
\arguments{
\item{x}{\code{sf} or \code{sfc}; Area used to select intersecting LIDAR tiles.}

\item{key}{\code{character}; LIDAR product to download. One of \code{"mnt"},
\code{"mns"} or \code{"mnh"}.}

\item{cache}{\code{character}; Cache directory. If \code{NULL}, the appropriate
Rsequoia2 LIDAR cache is used, see \code{\link[=seq_cache]{seq_cache()}}.}

\item{overwrite}{\code{logical}; If \code{TRUE}, overwrite existing files.}

\item{verbose}{\code{logical}; If \code{TRUE}, display progress and informational messages.}

\item{max_tries}{\code{integer}; Maximum number of download attempts.}
}
\value{
Invisibly returns a \code{character} vector of local tile paths.
}
\description{
Downloads IGN LIDAR HD tiles intersecting an area of interest.
}
\keyword{internal}
