% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.R
\name{extract_rows}
\alias{extract_rows}
\title{Extract all data rows from table}
\usage{
extract_rows(table_data, col_positions)
}
\arguments{
\item{table_data}{A \code{data.frame} from \code{\link[=group_lines_by_y]{group_lines_by_y()}}}

\item{col_positions}{List from \code{\link[=define_column_positions]{define_column_positions()}}}
}
\value{
List of \code{data.frame}, one per extracted row
}
\description{
Processes all lines in table data, extracting valid rows.
}
\keyword{internal}
