% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.R
\name{extract_visual_data}
\alias{extract_visual_data}
\title{Extract visual data from PDF with page numbers}
\usage{
extract_visual_data(pdf)
}
\arguments{
\item{pdf}{\code{character}. Path to PDF file}
}
\value{
A \code{data.frame} with columns from \code{pdf_data()} plus \code{page} (integer)
}
\description{
Wrapper around \code{\link[pdftools:pdftools]{pdftools::pdf_data()}} that adds page numbers
to each extracted text element.
}
\seealso{
\code{\link[pdftools:pdftools]{pdftools::pdf_data()}}.
}
\keyword{internal}
