% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessibility.R
\name{get_accessibility}
\alias{get_accessibility}
\title{Retrieve forest accessibility features (porter or skidder)}
\usage{
get_accessibility(
  x,
  type = c("porteur", "skidder"),
  buffer = 1000,
  verbose = TRUE
)
}
\arguments{
\item{x}{An `sf` object defining the input area of interest.}

\item{type}{`character` Accessibility type. One of `"porter"` or `"skidder"`.}

\item{buffer}{`numeric`; Buffer around `x` (in **meters**) used to enlarge
the download area.}

\item{verbose}{`logical` If `TRUE`, display messages.}
}
\value{
An `sf` object containing accessibility features, or `NULL` if none found.
}
\description{
Builds a convex buffer around the input geometry and retrieves accessibility
features for the requested machine type.
}
