% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geol.R
\name{get_brgm}
\alias{get_brgm}
\title{Download and read BRGM "BD Charm 50" harmonised geology layers}
\usage{
get_brgm(
  deps,
  source = "carhab",
  cache = NULL,
  verbose = FALSE,
  overwrite = FALSE
)
}
\arguments{
\item{deps}{\code{character} or \code{numeric}; One or sevral french
department code (see \link[happign:dep_2025]{happign::dep_2025}).}

\item{source}{\code{character} Source use to download geology from BRGM. Must be one of:
\itemize{
\item \code{"carhab"} : geological data used to created the \code{CarHab} dataset ;
\item \code{"bdcharm50"} : geological maps, vectorized and harmonised at 1:50,000 scale.
}}

\item{cache}{\code{character}; Storage directory. Defaults to the user cache
directory (see \code{\link[tools:userdir]{tools::R_user_dir()}}).}

\item{verbose}{\code{logical} If \code{TRUE}, display messages.}

\item{overwrite}{\code{logical} If \code{TRUE}, overwrite zipfile.}
}
\value{
An \code{sf} object containing the geology features for all requested
departments.

\code{sf}
}
\description{
Downloads and reads the geology layer (\code{S_FGEOL_*.shp}) from BRGM
"BD Charm 50" ZIP archives for one or several French
departments.
}
\details{
Multiple departments are combined into a single \code{sf} object.

For each department:
\itemize{
\item The corresponding BRGM ZIP archive is obtained via \code{\link[=download_brgm]{download_brgm()}} ;
\item The geology shapefile (\code{S_FGEOL_*.shp}) is imported ;
\item The layer is read directly from the ZIP using \code{sf::read_sf("/vsizip/...")}
without extracting files ;
}

The resulting layers are row-bound into a single \code{sf} object. All
geometries are returned in the CRS provided by BRGM (typically EPSG:2154).
}
