% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curves.R
\name{get_curves}
\alias{get_curves}
\title{Retrieve hypsometric curves around an area}
\usage{
get_curves(x, verbose = TRUE)
}
\arguments{
\item{x}{An \code{sf} object defining the input area of interest.}

\item{verbose}{\code{logical}; If \code{TRUE}, display progress and informational messages.}
}
\value{
An \code{sf} object containing hypsometric curves.
}
\description{
Builds a convex buffer around the input geometry, retrieves hypsometric
curves and returns an \code{sf} line layer.
}
\details{
The function creates a 1000 m convex buffer around the input geometry \code{x}
and retrieves hypsometric curves before returns as a single \code{sf} point layer.
}
