% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curves.R
\name{get_curves}
\alias{get_curves}
\title{Retrieve hypsometric curves around an area}
\usage{
get_curves(x, verbose = TRUE)
}
\arguments{
\item{x}{An `sf` object defining the input area of interest.}

\item{verbose}{`logical` If `TRUE`, display messages.}
}
\value{
An `sf` object containing hypsometric curves.
}
\description{
Builds a convex buffer around the input geometry, retrieves hypsometric
curves and returns an `sf` line layer.
}
\details{
The function creates a 1000 m convex buffer around the input geometry `x`
and retrieves hypsometric curves before returns as a single `sf` point layer.
}
