% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alti-rgealti.R
\name{get_dem}
\alias{get_dem}
\title{Download Digital Elevation Model (DEM) raster from IGN RGEAlti}
\usage{
get_dem(x, buffer = 200, res = 1, crs = 2154, verbose = TRUE)
}
\arguments{
\item{x}{`sf` or `sfc`; Geometry located in France.}

\item{buffer}{`numeric`; Buffer around `x` (in **meters**) used to enlarge
the download area.}

\item{res}{`numeric`; resolution specified in the units of the coordinate
system (see [happign::get_wms_raster()])}

\item{crs}{`numeric` or `character`; CRS of the returned raster (see
[happign::get_wms_raster()])}

\item{verbose}{`logical`; If `TRUE`, display messages.}
}
\value{
`SpatRaster` object from `terra` package
}
\description{
Downloads DEM from the IGN WMS service for the area covering `x` expanded
with a buffer.
The result is returned as a masked `SpatRaster`, clipped to the buffer
geometry to keep file size minimal.
}
\seealso{
[happign::get_wms_raster()]
}
