% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elevation.R
\name{get_dem}
\alias{get_dem}
\title{Download Digital Elevation Model (DEM) raster from IGN RGEAlti}
\usage{
get_dem(x, buffer = 200, res = 1, crs = 2154, verbose = TRUE)
}
\arguments{
\item{x}{\code{sf} or \code{sfc}; Geometry located in France.}

\item{buffer}{\code{numeric}; Buffer around \code{x} (in \strong{meters}) used to enlarge
the download area.}

\item{res}{\code{numeric}; resolution specified in the units of the coordinate
system (see \code{\link[happign:get_wms_raster]{happign::get_wms_raster()}})}

\item{crs}{\code{numeric} or \code{character}; CRS of the returned raster (see
\code{\link[happign:get_wms_raster]{happign::get_wms_raster()}})}

\item{verbose}{\code{logical}; If \code{TRUE}, display messages.}
}
\value{
\code{SpatRaster} object from \code{terra} package
}
\description{
Downloads DEM from the IGN WMS service for the area covering \code{x} expanded
with a buffer.
The result is returned as a masked \code{SpatRaster}, clipped to the buffer
geometry to keep file size minimal.
}
\seealso{
\code{\link[happign:get_wms_raster]{happign::get_wms_raster()}}
}
