% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geol.R
\name{get_geol}
\alias{get_geol}
\title{Download BRGM geology data to a geometry}
\usage{
get_geol(
  x,
  deps = NULL,
  source = "carhab",
  cache = NULL,
  verbose = FALSE,
  overwrite = FALSE
)
}
\arguments{
\item{x}{\code{sf} or \code{sfc}; Geometry used to determine relevant departments
and to spatially filter geology data.}

\item{deps}{\code{character} or \code{numeric}; One or sevral french
department code (see \code{\link[=get_cog]{get_cog()}}).}

\item{source}{\code{character} Source use to download geology from BRGM. Must be one of:
\itemize{
\item \code{"carhab"} : geological data used to created the \code{CarHab} dataset ;
\item \code{"bdcharm50"} : geological maps, vectorized and harmonised at 1:50,000 scale.
}}

\item{cache}{\code{character}; Storage directory. Defaults to the user cache
directory (see \code{\link[tools:userdir]{tools::R_user_dir()}}).}

\item{verbose}{\code{logical} If \code{TRUE}, display messages.}

\item{overwrite}{\code{logical} If \code{TRUE}, overwrite zipfile.}

\item{buffer}{\code{numeric}; Buffer distance (in meters) applied to \code{x}
before spatial filtering. Default is \code{100}.}
}
\value{
An \code{sf} object containing clipped geology features.
}
\description{
Downloads BRGM geology layers for the departments intersecting \code{x}
(or for specified departments) and spatially filters them using a
buffered envelope around \code{x}.
}
\details{
The function:
\itemize{
\item Identifies departments intersecting \code{x} (if \code{deps} is not provided) ;
\item Downloads geology layers using \code{\link[=get_brgm]{get_brgm()}} ;
\item Filters features intersecting a buffered envelope around \code{x}.
}

All geometries are returned in EPSG:2154.
}
