% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpu.R
\name{get_gpu}
\alias{get_gpu}
\title{Download GPU vector layer}
\usage{
get_gpu(x, key, verbose = TRUE)
}
\arguments{
\item{x}{`sf` or `sfc`; Geometry located in France.}

\item{key}{`character`; Layer to download.
Must be one of from `get_keys("gpu", reduce = FALSE)`}

\item{verbose}{`logical`; If `TRUE`, display messages.}
}
\value{
`sf` object from `sf` package
}
\description{
Downloads a vector layer with `hhapign` for the area covering `x`
expanded with a buffer.
}
