% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpu.R
\name{get_gpu}
\alias{get_gpu}
\title{Download GPU vector layer}
\usage{
get_gpu(x, key, verbose = TRUE)
}
\arguments{
\item{x}{\code{sf} or \code{sfc}; Geometry located in France.}

\item{key}{\code{character}; Layer to download.
Must be one of from \code{get_keys("gpu", reduce = FALSE)}}

\item{verbose}{\code{logical}; If \code{TRUE}, display messages.}
}
\value{
\code{sf} object from \code{sf} package
}
\description{
Downloads a vector layer with \code{hhapign} for the area covering \code{x}
expanded with a buffer.
}
