% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro.R
\name{get_hydro_poly}
\alias{get_hydro_poly}
\title{Retrieve hydrographic polygons around an area}
\usage{
get_hydro_poly(x, buffer = 1000)
}
\arguments{
\item{x}{An \code{sf} object used as the input area.}

\item{buffer}{\code{numeric}; Buffer around \code{x} (in \strong{meters}) used to enlarge}
}
\value{
An \code{sf} object containing hydrographic polygons with two fields:
\itemize{
\item \code{TYPE} - hydrographic class
\itemize{
\item \code{RSO} = Reservoir or water tower
\item \code{SFP} = Permanent hydrographic surface
\item \code{SFI} = Intermittent hydrographic surface
}
\item \code{NATURE} - Original BDTOPO nature field
\item \code{NAME} - Official hydrographic name (when available)
}
}
\description{
Retrieve hydrographic polygons around an area
}
\details{
The function retrieves BDTOPO layers within a convex buffer
around \code{x}, assigns the hydrographic types, and combines them into
a single layer.
}
