% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legal_entity.R
\name{get_legal_entity}
\alias{get_legal_entity}
\title{Create a forest matrice for legal entity from insee code}
\usage{
get_legal_entity(x, cache = NULL, verbose = TRUE)
}
\arguments{
\item{x}{\code{character}; Code(s) INSEE or code(s) department (see
\link[happign:com_2025]{happign::com_2025} or \link[happign:dep_2025]{happign::dep_2025})}

\item{cache}{\code{character}; Storage directory. Defaults to the user cache
directory (see \code{\link[tools:userdir]{tools::R_user_dir()}}).}

\item{verbose}{\code{logical} If \code{TRUE}, display messages.}
}
\description{
Generates a forest matrice used to store general forest
information (e.g., \code{IDENTIFIANT}, \code{PROPRIETAIRE}) and cadastral attributes
(\code{CODE_INSEE}, \code{PREFIXE}, \code{SECTION}, \code{NUMERO}, \code{LIEU_DIT}) for legal
entity.
}
\details{
Matrice from \code{get_legal_entity} add \code{tx_boisee} information which is the
percentage of land considered as forest :
\itemize{
\item \code{"L"}:  Landes
\item \code{"LB"}: Landes Boisées
\item \code{"B"}:  Bois
\item \code{"BF"}: Futaies Feuillues
\item \code{"BM"}: Futaies Mixtes
\item \code{"BO"}: Oseraies
\item \code{"BP"}: Peupleraies
\item \code{"BR"}: Futaies Résineuses
\item \code{"BS"}: Taillis sous Futaies
\item \code{"BT"}: Taillis Simples
}
}
\examples{
\dontrun{
pm <- get_legal_entity(c("29158", "08185"))
}

}
