% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.R
\name{get_matrice}
\alias{get_matrice}
\title{Process multiple PDFs and combine results}
\usage{
get_matrice(pdf)
}
\arguments{
\item{pdf}{\code{character} vector. Paths to PDF file}
}
\value{
List with three \code{data.frame}: \code{refs}, \code{owners}, \code{parcels}
}
\description{
Applies the extraction pipeline to multiple PDF files and combines
results into a single list with three data frames.
}
