% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnhn.R
\name{get_mnhn}
\alias{get_mnhn}
\title{Download MNHN vector layer from IGN WFS (RGB or IRC)}
\usage{
get_mnhn(x, key, buffer = 500, verbose = TRUE)
}
\arguments{
\item{x}{\code{sf} or \code{sfc}; Geometry located in France.}

\item{key}{\code{character}; Layer to download. Must be one of from \code{get_keys("mnhn")}}

\item{buffer}{\code{numeric}; Buffer around \code{x} (in \strong{meters}) used to enlarge
the download area.}

\item{verbose}{\code{logical}; If \code{TRUE}, display messages.}
}
\value{
\code{sf} object from \code{sf} package
}
\description{
Downloads a vector layer from the IGN WFS service for the
area covering \code{x} expanded with a buffer.
Available layer are from MNHN (Nationale Museum of Natural History)
}
\examples{
\dontrun{
library(sf)

x <- st_sfc(st_point(c(-4.372746579180652, 47.79820761331345)), crs = 4326)

ospar <- get_mnhn(x, key = "ospar", buffer = 5000)
x <- st_transform(x, st_crs(ospar))

plot(st_geometry(ospar), col = "lightgrey")
plot(st_geometry(st_buffer(x, 5000)), col = "#FFB3B3", add = TRUE)
plot(st_geometry(x), col = "red", lwd =2, add = TRUE, pch = 19, cex = 2)

}
}
\seealso{
\code{\link[happign:get_wfs]{happign::get_wfs()}}
}
