% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old.R
\name{get_old}
\alias{get_old}
\title{Retrieve OLD features around an area}
\usage{
get_old(x, buffer = 1000, verbose = TRUE)
}
\arguments{
\item{x}{An `sf` object defining the input area of interest.}

\item{buffer}{`numeric`; Buffer around `x` (in **meters**) used to enlarge}

\item{verbose}{`logical` If `TRUE`, display messages.}
}
\value{
An `sf` object containing OLD features.
}
\description{
Builds a convex buffer around the input geometry, retrieves OLD
features and returns an `sf` point layer.
}
\details{
The function creates a convex buffer around the input geometry `x`
and retrieves OLD features before returns as a single `sf` point layer.
}
