% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ortho.R
\name{get_ortho}
\alias{get_ortho}
\title{Download orthophotos from the IGN WMS (RGB or IRC)}
\usage{
get_ortho(
  x,
  type = c("irc", "rgb"),
  buffer = 200,
  zoom = 12,
  crs = 2154,
  overwrite = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{x}{`sf` or `sfc`; Geometry located in France.}

\item{type}{`character`; Type of orthophoto to download. Must be one of:
- `"rgb"` - true-color orthophoto
- `"irc"` - near-infrared orthophoto}

\item{buffer}{`numeric`; Buffer around `x` (in **meters**) used to enlarge
the download area.}

\item{zoom}{`integer` between 0 and 21. The smaller the zoom level, the less
precise the resolution(see [happign::get_wmts()])}

\item{crs}{`numeric` or `character`; CRS of the returned raster (see
[happign::get_wmts()])}

\item{overwrite}{`logical`; If `TRUE`, file is overwritten.}

\item{verbose}{`logical`; If `TRUE`, display messages.}
}
\value{
`SpatRaster` object from `terra` package
}
\description{
Downloads an orthophoto (RGB or infrared/IRC) from the IGN WMTS service for the
area covering `x` expanded with a buffer.
The result is returned as a masked `SpatRaster`, clipped to the input geometry
to keep file size minimal.
}
\details{
The orthophoto retrieved contains data for the whole bounding
box (bbox) of `x` (plus the buffer).
To reduce the final file size and avoid unnecessary pixels, the raster is
immediately masked with the buffered geometry.
}
\examples{
\dontrun{

p <- sf::st_sfc(st_point(c(-4.372746579180652, 47.79820761331345)), crs = 4326)

ortho <- get_ortho(p, type = "rgb", buffer = 50)
irc <- get_ortho(p, type = "irc", buffer = 50)

terra::plotRGB(ortho)
terra::plotRGB(irc)

}

}
\seealso{
[happign::get_wmts()]
}
