% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parca.R
\name{get_parca_etalab}
\alias{get_parca_etalab}
\title{Retrieve a cadastral parcel geometry from Etalab}
\usage{
get_parca_etalab(idu)
}
\arguments{
\item{idu}{`character` Cadastral parcel identifier.}
}
\value{
An `sf` object containing the parcel geometry.
}
\description{
Retrieve a cadastral parcel geometry from Etalab
}
