% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parca.R
\name{get_parca_etalab}
\alias{get_parca_etalab}
\title{Retrieve a cadastral parcel geometry from Etalab}
\usage{
get_parca_etalab(idu, verbose = TRUE)
}
\arguments{
\item{idu}{\code{character} Cadastral parcel identifier.}

\item{verbose}{\code{logical} If \code{TRUE}, display progress messages.}
}
\value{
An \code{sf} object containing the parcel geometry.
}
\description{
Retrieve a cadastral parcel geometry from Etalab
}
