% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patrimony.R
\name{get_patrimony}
\alias{get_patrimony}
\title{Download patrimony vector layer}
\usage{
get_patrimony(x, key, buffer = 500, verbose = TRUE)
}
\arguments{
\item{x}{`sf` or `sfc`; Geometry located in France.}

\item{key}{`character`; Layer to download.
Must be one of from `get_keys("pat")`}

\item{buffer}{`numeric`; Buffer around `x` (in **meters**) used to enlarge
the download area.}

\item{verbose}{`logical`; If `TRUE`, display messages.}
}
\value{
`sf` object from `sf` package
}
\description{
Downloads a vector layer with `frheritage` for the area covering `x`
expanded with a buffer.
}
