% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedology.R
\name{get_pedology}
\alias{get_pedology}
\title{Retrieve pedology polygon features around an area}
\usage{
get_pedology(x)
}
\arguments{
\item{x}{An \code{sf} object defining the input area of interest.}
}
\value{
An \code{sf} object containing pedology polygon features
intersecting the input geometry, with additional surface fields.
}
\description{
Retrieves pedological polygon features from the INRA soil map
intersecting an area of interest and computes surface attributes.
}
\details{
The function retrieves pedology polygon features from the
\code{INRA.CARTE.SOLS:geoportail_vf} layer intersecting the input geometry \code{x}.
The resulting geometries are intersected with \code{x}, cast to polygons,
and surface attributes are computed using \code{ua_generate_area()}.
}
\seealso{
\code{\link[=ua_generate_area]{ua_generate_area()}}
}
