% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prsf.R
\name{get_prsf}
\alias{get_prsf}
\title{Retrieve PRSF point features around an area}
\usage{
get_prsf(x, buffer = 5000, verbose = TRUE)
}
\arguments{
\item{x}{An `sf` object defining the input area of interest.}

\item{buffer}{`numeric`; Buffer around `x` (in **meters**) used to enlarge}

\item{verbose}{`logical` If `TRUE`, display messages.}
}
\value{
An `sf` object containing PRSF point features.
}
\description{
Builds a convex buffer around the input geometry, retrieves PRSF
point features and returns an `sf` point layer.
}
\details{
The function creates convex buffer around the input geometry `x`
and retrieves PRSF point features before returns as a single `sf` point layer.
}
