% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prsf.R
\name{get_prsf}
\alias{get_prsf}
\title{Retrieve PRSF point features around an area}
\usage{
get_prsf(x)
}
\arguments{
\item{x}{An \code{sf} object defining the input area of interest.}
}
\value{
An \code{sf} object containing PRSF point features.
}
\description{
Builds a convex buffer around the input geometry, retrieves PRSF
point features and returns an \code{sf} point layer.
}
\details{
The function creates a 5000 m convex buffer around the input geometry \code{x}
and retrieves PRSF point features before returns as a single \code{sf} point layer.
}
