% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.R
\name{get_raw_parcels}
\alias{get_raw_parcels}
\title{Extract raw parcel data from PDF}
\usage{
get_raw_parcels(pdf)
}
\arguments{
\item{pdf}{\code{character}. Path to PDF file}
}
\value{
A \code{data.frame} with raw parcel data, column names from
\code{\link[=define_column_positions]{define_column_positions()}}
}
\description{
Main function that orchestrates the extraction of unbuilt parcel data
from French cadastral PDFs. Calls all internal processing functions.
}
\keyword{internal}
