% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elevation.R
\name{get_slope}
\alias{get_slope}
\title{Compute a Slope Raster from a DEM}
\usage{
get_slope(x = NULL, dem = NULL, agg = 5, verbose = TRUE, ...)
}
\arguments{
\item{x}{\code{sf} or \code{sfc}; Geometry located in France.}

\item{dem}{\code{SpatRaster} representing ground elevation (DEM). Must be supplied
only when \code{x} is \code{NULL}.}

\item{agg}{\code{numeric}; Target resolution (in meters) to which the DEM is
aggregated if its native resolution is finer. Default: \code{5}.}

\item{verbose}{\code{logical}; If \code{TRUE}, display messages.}

\item{...}{Additional parameters passed to \code{\link[=get_dem]{get_dem()}} when \code{x} is supplied.}
}
\value{
A \code{SpatRaster} containing slope values in degrees.
}
\description{
Computes a slope raster (in degrees) either by:
\itemize{
\item downloading a DEM automatically using \code{x}, or
\item using a DEM raster supplied manually through the \code{dem} argument.
}
}
\details{
When \code{x} is provided, the DEM is obtained with \code{\link[=get_dem]{get_dem()}}, using by
default a resolution equal to \code{agg}.
When \code{x} is not provided, a DEM must be supplied via \code{dem}.

If the DEM resolution is finer than \code{agg}, the raster is aggregated to
avoid artefacts and reduce computation time.

Slope is computed with \code{terra::terrain()} using Horn's 8-neighbor algorithm,
and returned in \strong{degrees}.

Aggregation is performed with \code{terra::aggregate()} using a mean function.
This is recommended when high-resolution DEMs (<5 m) would otherwise
generate line artefacts.
}
\examples{
\dontrun{
# Automatic download mode
s <- get_slope(x = my_polygon, buffer = 200)

# Manual mode
dem <- get_dem(my_polygon)
s <- get_slope(dem = dem)
}
}
