% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vege.R
\name{get_vege_line}
\alias{get_vege_line}
\title{Retrieve forest vegetation lines around an area}
\usage{
get_vege_line(x)
}
\arguments{
\item{x}{`sf` or `sfc`; Geometry located in France.}
}
\value{
An `sf` object containing forest vegetation line features.
}
\description{
Retrieve forest vegetation lines around an area
}
\details{
The function derives forest vegetation linear features from
vegetation polygons obtained with [get_vege_poly()].

GIS workflow :
- Polygon are intersected to a 1500m buffer;
- Polygon are cast to `LINESTRING`;
- Line are intersected with a 1500 buffer to remove intersection line
from first step

If no vegetation polygons are available, the function returns
an empty standardized `sf` object.
}
\seealso{
get_vege_poly
}
