% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vege.R
\name{get_vege_poly}
\alias{get_vege_poly}
\title{Retrieve forest vegetation polygons around an area}
\usage{
get_vege_poly(x)
}
\arguments{
\item{x}{\code{sf} or \code{sfc}; Geometry located in France.}
}
\value{
An \code{sf} object containing forest vegetation polygons
}
\description{
Retrieve forest vegetation polygons around an area
}
\details{
The function retrieves the IGN forest mask layer within a 1000 m
convex buffer around \code{x}. Retrieved polygons are then intersected
with a 1500 m convex buffer, cast to polygons, and normalized.

If no forest data are found, the function returns an empty
\code{sf} object with a standardized structure.
}
