% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{idu_build}
\alias{idu_build}
\title{Build IDU}
\usage{
idu_build(dep, com, prefix, section, numero)
}
\arguments{
\item{idu}{\code{character} IDU(s) to pad}
}
\description{
Internal helper used to build idu
}
\keyword{internal}
