% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteo-france.R
\name{mf_get_nearest_station}
\alias{mf_get_nearest_station}
\title{Get nearest Meteo-France stations}
\usage{
mf_get_nearest_station(x, n = 1, verbose = TRUE)
}
\arguments{
\item{x}{`sf` or `sfc`.}

\item{n}{`integer`; number of nearest stations to return. Default is 1.}

\item{verbose}{`logical` If `TRUE`, display progress messages.}
}
\value{
An `sf` object containing the nearest meteorological stations.
}
\description{
Finds the nearest meteorological stations from Meteo-France to the provided
spatial geometry.
}
