% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteo-france.R
\name{mf_get_nearest_station}
\alias{mf_get_nearest_station}
\title{Get nearest Météo-France stations}
\usage{
mf_get_nearest_station(x, n = 1, verbose = TRUE)
}
\arguments{
\item{x}{\code{sf} or \code{sfc}.}

\item{n}{\code{integer}; number of nearest stations to return. Default is 1.}

\item{verbose}{\code{logical} If \code{TRUE}, display progress messages.}
}
\value{
An \code{sf} object containing the nearest meteorological stations.
}
\description{
Finds the nearest meteorological stations from Météo-France to the provided
spatial geometry.
}
