% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteo-france.R
\name{mf_ombro}
\alias{mf_ombro}
\title{Compute ombrothermic climatology summaries}
\usage{
mf_ombro(clim, periods = c(30, 5))
}
\arguments{
\item{clim}{\code{data.frame} Raw climatology data from \code{\link[=mf_get_climatology]{mf_get_climatology()}}}

\item{periods}{\code{integer} Vector of periods (in years) used to compute averages.
Default is \code{c(30, 5)}.}
}
\value{
A \code{data.frame} containing monthly averages of temperature and precipitation
for each period.
}
\description{
Computes monthly ombrothermic summaries from raw climatology data over
specified rolling periods.
}
\examples{
\dontrun{
ombro <- mf_ombro(clim)
}

}
