% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parca.R
\name{parca_check_area}
\alias{parca_check_area}
\title{Check inconsistencies between cadastral and cartographic areas}
\usage{
parca_check_area(parca, atol = 500, rtol = 0.05, verbose = TRUE)
}
\arguments{
\item{parca}{`sf` Object from [Rsequoia2::seq_parca()] representing cadastral
parcels.}

\item{atol}{`numeric` Absolute difference tolerance in m2. Default to `500m`.}

\item{rtol}{`numeric` Relative difference. Default to `0.05`}

\item{verbose}{`logical` If `TRUE`, display messages.}
}
\value{
The input `parca` with four additional fields:
  `AREA_SIG` (cartographic area in ha),
  `ATOL_AREA` (absolute difference in m2),
  `RTOL_AREA` (relative difference),
  `CHECK_AREA` (logical flag).
}
\description{
This function compares cadastral areas (in m2) with
cartographic areas computed from geometry ([sf::st_area()]).
}
