% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parca.R
\name{parca_check_area}
\alias{parca_check_area}
\title{Check inconsistencies between cadastral and cartographic areas}
\usage{
parca_check_area(parca, atol = 500, rtol = 0.05, verbose = TRUE)
}
\arguments{
\item{parca}{\code{sf} Object from \code{\link[=seq_parca]{seq_parca()}} representing cadastral
parcels.}

\item{atol}{\code{numeric} Absolute difference tolerance in m². Default to \verb{500m}.}

\item{rtol}{\code{numeric} Relative difference. Default to \code{0.05}}

\item{verbose}{\code{logical} If \code{TRUE}, display messages.}
}
\value{
The input \code{parca} with four additional fields:
\code{AREA_SIG} (cartographic area in ha),
\code{ATOL_AREA} (absolute difference in m²),
\code{RTOL_AREA} (relative difference),
\code{CHECK_AREA} (logical flag).
}
\description{
This function compares cadastral areas (in m²) with
cartographic areas computed from geometry (\code{\link[sf:geos_measures]{sf::st_area()}}).
}
