% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ua.R
\name{parca_to_ua}
\alias{parca_to_ua}
\title{Create \emph{UA} sf object from \emph{PARCA} sf object}
\usage{
parca_to_ua(parca)
}
\arguments{
\item{parca}{\code{sf} Object from \code{\link[=seq_parca]{seq_parca()}} containing cadastral
parcels.}
}
\value{
An \code{sf} object containing the analysis units.
}
\description{
Create \emph{UA} sf object from \emph{PARCA} sf object
}
