% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legal_entity.R
\name{search_legal_entity}
\alias{search_legal_entity}
\title{Helpers to search within a forest matrix}
\usage{
search_legal_entity(matrice, prop = NULL, lieu_dit = NULL)
}
\arguments{
\item{matrice}{\code{data.frame}; The matrix containing parcel information in a
format readable by SEQUOIA.}

\item{prop}{\code{character}; Owner name(s) used to further filter results. Defaults to \code{NULL}.
Multiple owners can be specified.}

\item{lieu_dit}{\code{character}; Locality name(s) used to further filter results. Defaults to \code{NULL}.
Multiple locality names can be specified.}
}
\value{
A \code{data.frame}.
}
\description{
This function allows searching for parcels in a forest matrix based on
owner names or location (lieu-dit).
}
\details{
The search relies on a text–normalization step applied to both the inputs and
the corresponding columns of the matrice. This makes the search robust
to accents, punctuation, spacing irregularities and case differences.

\strong{Examples of normalization:}
\itemize{
\item \code{"État / Forêts"} → \code{"ETATFORETS"}
\item \code{"  Le Bois-de l'Orme  "} → \code{"LEBOISDELORME"}
\item \code{"Société du Chêne"} → \code{"SOCIETEDUCHENE"}
}
}
\examples{
\dontrun{
matrice <- get_legal_entity(29158)
search_mat <- search_legal_entity(matrice, c("penmarch", "guenole"))
}

}
