% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{seq1_id}
\alias{seq1_id}
\title{Retrieve old `R_SEQUOIA` _PARCA_ identifier}
\usage{
seq1_id(dirname = ".")
}
\arguments{
\item{dirname}{`character` Directory where the file _PARCA_ is located.
Defaults to the current working directory.}
}
\value{
The `character` identifier
}
\description{
Retrieve old `R_SEQUOIA` _PARCA_ identifier
}
