% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{seq1_id}
\alias{seq1_id}
\title{Retrieve old \code{R_SEQUOIA} \emph{PARCA} identifier}
\usage{
seq1_id(dirname = ".")
}
\arguments{
\item{dirname}{\code{character} Directory where the file \emph{PARCA} is located.
Defaults to the current working directory.}
}
\value{
The \code{character} identifier
}
\description{
Retrieve old \code{R_SEQUOIA} \emph{PARCA} identifier
}
