% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{seq1_read}
\alias{seq1_read}
\title{Open old `R_SEQUOIA` vector file}
\usage{
seq1_read(dirname = ".", layer)
}
\arguments{
\item{dirname}{`character` Directory where the file is located.
Defaults to the current working directory.}

\item{layer}{`character` Layer name.}
}
\value{
An `sf` object or an empty `sf` object if no file are matched.
}
\description{
Open old `R_SEQUOIA` vector file
}
