% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{seq1_read}
\alias{seq1_read}
\title{Open old \code{R_SEQUOIA} vector file}
\usage{
seq1_read(dirname = ".", layer)
}
\arguments{
\item{dirname}{\code{character} Directory where the file is located.
Defaults to the current working directory.}

\item{layer}{\code{character} Layer name.}
}
\value{
An \code{sf} object.
}
\description{
Open old \code{R_SEQUOIA} vector file
}
