% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{seq1_update}
\alias{seq1_update}
\title{Update `R_SEQUOIA` projet to `Rsequoia2`}
\usage{
seq1_update(dirname = ".", verbose = TRUE, overwrite = FALSE)
}
\arguments{
\item{dirname}{`character` Directory where the `R_SEQUOIA` files are located.
Defaults to the current working directory.}

\item{verbose}{`logical` If `TRUE`, display progress messages.
Defaults to `TRUE`.}

\item{overwrite}{`logical`; whether to overwrite existing files.
Defaults to `FALSE`.}
}
\value{
Invisibly returns a named vector of file paths written by [seq_write()].
Returns `NULL` invisibly when no layer are found.
}
\description{
Update `R_SEQUOIA` projet to `Rsequoia2`
}
