% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessibility.R
\name{seq_access}
\alias{seq_access}
\title{Generate access point layer for a Sequoia project}
\usage{
seq_access(dirname = ".", verbose = TRUE, overwrite = FALSE)
}
\arguments{
\item{dirname}{`character` Path to the project directory.
Defaults to the current working directory.}

\item{verbose}{`logical`; whether to display informational messages.
Defaults to `TRUE`.}

\item{overwrite}{`logical`; whether to overwrite existing files.
Defaults to `FALSE`.}
}
\value{
Invisibly returns a named list of file paths written by [seq_write()].
}
\description{
Create an empty `sf` for access point features, and writes the resulting
layer to disk.
}
\details{
The access point layer is an empty layer : user must point access themselves.
The layer is written to disk using [seq_write()] with the key `"v.access.entry.point"`.
}
\seealso{
[seq_write()]
}
