% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\name{seq_access}
\alias{seq_access}
\title{Generate access point layer for a Sequoia project}
\usage{
seq_access(dirname = ".", verbose = TRUE, overwrite = FALSE)
}
\arguments{
\item{dirname}{\code{character} Path to the project directory.
Defaults to the current working directory.}

\item{verbose}{\code{logical}; whether to display informational messages.
Defaults to \code{TRUE}.}

\item{overwrite}{\code{logical}; whether to overwrite existing files.
Defaults to \code{FALSE}.}
}
\value{
Invisibly returns a named list of file paths written by \code{\link[=seq_write]{seq_write()}}.
}
\description{
Create an empty \code{sf} for access point features, and writes the resulting
layer to disk.
}
\details{
The access point layer is an empty layer : user must point access themselves.
The layer is written to disk using \code{\link[=seq_write]{seq_write()}} with the key \code{"v.acces.point"}.
}
\seealso{
\code{\link[=seq_write]{seq_write()}}
}
