% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-cache.R
\name{seq_cache_clear}
\alias{seq_cache_clear}
\title{Clear Rsequoia2 cache}
\usage{
seq_cache_clear(key = NULL, filepath = NULL, ask = TRUE)
}
\arguments{
\item{key}{`character`. Cache key to clear. If `NULL`, all configured caches
are cleared.}

\item{filepath}{`character`. Optional path to a cache configuration file.
Mostly useful for tests.}

\item{ask}{`logical`. Ask confirmation before deleting cache files.}
}
\value{
Invisibly returns `TRUE`.
}
\description{
Deletes one or all configured Rsequoia2 cache folders after user confirmation.
}
