% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{seq_field}
\alias{seq_field}
\title{Load field definitions from the Sequoia configuration}
\usage{
seq_field(field = NULL, filepath = NULL)
}
\arguments{
\item{field}{\code{character} or \code{NULL}; If provided, must match a key in
\code{inst/config/seq_fields.yaml}.}

\item{filepath}{\code{character} or \code{NULL}; override for the path to the default
YAML configuration file (\code{inst/config/seq_fields.yaml}). Used mainly for
testing.}
}
\value{
A list describing all fields, or a single field definition.
}
\description{
Internal helper used by \code{seq_normalize()} to access field metadata
defined in \code{inst/config/seq_fields.yaml}. When \code{field} is \code{NULL}, the full
configuration list is returned. Otherwise, the corresponding field
definition is extracted.
}
\examples{
\dontrun{
# Return all field definitions
seq_field()

# All avaialbe fields
names(seq_field())

# Return the definition of a specific field
seq_field("identifiant")
}

}
