% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geol.R
\name{seq_geol}
\alias{seq_geol}
\title{Create geology layers for a Sequoia project from BRGM data}
\usage{
seq_geol(
  dirname = ".",
  key = NULL,
  cache = NULL,
  verbose = TRUE,
  overwrite = FALSE
)
}
\arguments{
\item{dirname}{\code{character} Directory where the matrice file is located.
Defaults to the current working directory.}

\item{key}{\code{character}. Optional geology layer identifier(s). If \code{NULL}
(default), all available geology layers are created. Available layers
are \code{"bdcharm50"} and \code{"carhab"}. Partial matching is supported
(see \code{\link[=seq_key]{seq_key()}}).}

\item{cache}{\code{character}; Storage directory. Defaults to the user cache
directory (see \code{\link[tools:userdir]{tools::R_user_dir()}}).}

\item{verbose}{\code{logical} If \code{TRUE}, display messages.}

\item{overwrite}{\code{logical} If \code{TRUE}, overwrite zipfile.}
}
\value{
An invisible named \code{list} of file paths to the created layers.
}
\description{
Uses the project's \emph{PARCA} layer to download geology datasets and clip
them to its geometry.
}
\details{
Layers and their associated \emph{QML style files} are written to the
project directory using \code{\link[=seq_write]{seq_write()}}.

\strong{Difference between BRGM and CARHAB data}

The BRGM 1:50000 geological maps are detailed, heterogeneous products
created for geological interpretation. They include many finely split
geological units and may show inconsistencies between neighbouring map
sheets due to varying survey dates, methods, and levels of detail.

CARHAB data are a simplified, harmonised reinterpretation of these maps.
Geological formations are recoded into broader lithological classes tailored
for ecological modelling. The goal is to provide consistent, comparable,
and ecologically relevant information across departments.

More info at \href{https://infoterre.brgm.fr/page/carhab-donnees-geologiques}{infoterre}
}
