% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-layers.R
\name{seq_layer}
\alias{seq_layer}
\title{Load layer config from the Sequoia configuration}
\usage{
seq_layer(pattern = NULL, filepath = NULL)
}
\arguments{
\item{pattern}{\code{character} or \code{NULL}; regular-expression pattern used to
filter keys. If \code{NULL}, all keys defined in the configuration are returned.}

\item{filepath}{\code{character} or \code{NULL}; optional override for the path to the
YAML configuration file. Mainly used for testing purposes.}
}
\value{
A list describing matched layer
}
\description{
Internal helper used to determine layer info like name or extension are stored
in \code{inst/config/seq_layerss.yaml}.
}
\examples{
\dontrun{
# List all available tables
names(seq_table())

# Load field keys for the "parcelle" table
seq_table("parca")
}

}
