% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alti-lidar.R
\name{seq_lidar}
\alias{seq_lidar}
\title{Create LiDAR layers for a Sequoia project}
\usage{
seq_lidar(
  dirname = ".",
  key = c("mnt", "mns", "mnh"),
  buffer = 200,
  crs = 2154,
  cache = NULL,
  overwrite = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{dirname}{\code{character} Directory where the matrice file is located.
Defaults to the current working directory.}

\item{key}{\code{character}; LiDAR product(s) to create. One or more of \code{"mnt"},
\code{"mns"} and \code{"mnh"}.}

\item{buffer}{\code{numeric}; Buffer distance, in meters, applied around \code{x}
before masking.}

\item{crs}{\code{integer} or \code{character}; Target CRS of the returned raster.
Defaults to EPSG:2154.}

\item{cache}{\code{character}; Cache directory. If \code{NULL}, the appropriate
Rsequoia2 LIDAR cache is used, see \code{\link[=seq_cache]{seq_cache()}}.}

\item{overwrite}{\code{logical}; If \code{TRUE}, re-download existing tiles.}

\item{verbose}{\code{logical}; If \code{TRUE}, display messages.}
}
\value{
Invisibly returns a named \code{list} of created raster file paths.
}
\description{
Uses the project's \emph{PARCA} layer to download, extract, and write LiDAR HD
altimetry rasters.
}
