% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alti-lidar.R
\name{seq_lidar}
\alias{seq_lidar}
\title{Create LiDAR layers for a Sequoia project}
\usage{
seq_lidar(
  dirname = ".",
  key = c("mnt", "mnh"),
  buffer = 200,
  crs = 2154,
  cache = NULL,
  overwrite = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{dirname}{`character` Directory where the matrice file is located.
Defaults to the current working directory.}

\item{key}{`character`; LiDAR product(s) to create. One or more of `"mnt"`,
`"mns"` and `"mnh"`.}

\item{buffer}{`numeric`; Buffer distance, in meters, applied around `x`
before masking.}

\item{crs}{`integer` or `character`; Target CRS of the returned raster.
Defaults to EPSG:2154.}

\item{cache}{`character`; Cache directory. If `NULL`, the appropriate
Rsequoia2 LIDAR cache is used, see [Rsequoia2::seq_cache()].}

\item{overwrite}{`logical`; If `TRUE`, re-download existing tiles.}

\item{verbose}{`logical`; If `TRUE`, display messages.}
}
\value{
Invisibly returns a named `list` of created raster file paths.
}
\description{
Uses the project's _PARCA_ layer to download, extract, and write LiDAR HD
altimetry rasters.
}
