% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteo-france.R
\name{seq_meteo_france}
\alias{seq_meteo_france}
\title{Download Meteo-France meteorological data for a sequoia process}
\usage{
seq_meteo_france(
  dirname = ".",
  cache = NULL,
  verbose = TRUE,
  overwrite = FALSE
)
}
\arguments{
\item{dirname}{`character` Directory where the matrice file is located.
Defaults to the current working directory.}

\item{cache}{`character`; Storage directory. Defaults to the user cache
directory (see [tools::R_user_dir()])}

\item{verbose}{`logical` If `TRUE`, display messages.}

\item{overwrite}{`logical` If `TRUE`, file is overwritten.}
}
\value{
A named `list` of file paths :
- `"fiche.meteo"``: Path to the downloaded climatological station PDF
- `"meteofrance"``: Path to the generated Excel workbook
}
\description{
Downloads and formats meteorological data from Meteo-France.
This includes download of the nearest climatological station fiche (PDF),
ombrothermic summaries, precipitation statistics. Results are written
to an Excel workbook based on the internal template `CLIMAT_MF.xlsx`.
}
