% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnhn.R
\name{seq_mnhn}
\alias{seq_mnhn}
\title{Download MNHN vector layers for a Sequoia project}
\usage{
seq_mnhn(
  dirname = ".",
  buffer = 500,
  key = get_keys("mnhn"),
  verbose = TRUE,
  overwrite = FALSE
)
}
\arguments{
\item{dirname}{`character` Directory where the matrice file is located.
Defaults to the current working directory.}

\item{buffer}{`numeric`; Buffer around `x` (in **meters**) used to enlarge
the download area.}

\item{key}{`character`; List of MNHN layer identifiers to download. If not
provided, the function uses `get_keys("mnhn")` to automatically select all
MNHN layers defined in the Sequoia configuration (`inst/config/seq_layers.yaml`)}

\item{verbose}{`logical`; If `TRUE`, display messages.}

\item{overwrite}{`logical` If `TRUE`, file is overwritten.}
}
\value{
A named list of file paths written by [seq_write()], one per layer.
}
\description{
Downloads one or several vector layers from the IGN WFS service
for `mnhn` layer(s) of a Sequoia project.
}
\details{
This function is a convenience wrapper looping over [get_mnhn()], allowing
the user to download all products in one call and automatically write them
to the project directory using [seq_write()].


For each value in `key`, the function attempts to query the corresponding
MNHN layer using [get_mnhn()].

- If the layer contains features, it is written to the project directory
  via [seq_write()] and recorded as a successful download.
- If the layer contains no features, it is skipped and marked
  as empty.
}
\seealso{
[get_mnhn()], [seq_write()]
}
