% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ua-occupation.R
\name{seq_occupation}
\alias{seq_occupation}
\title{Create occupation layer from _UA_ for a Sequoia project}
\usage{
seq_occupation(dirname = ".", verbose = TRUE, overwrite = FALSE)
}
\arguments{
\item{dirname}{`character` Directory where the matrice file is located.
Defaults to the current working directory.}

\item{verbose}{`logical` If `TRUE`, display messages.}

\item{overwrite}{`logical` If `TRUE`, file is overwritten.}
}
\value{
An invisible file path to the exported occupation layer.
}
\description{
Reads the UA polygon layer, aggregates surfaces by occupation status, and
writes the derived occupation layer to the project directory.
}
\seealso{
[ua_to_occupation()], [seq_read()], [seq_write()]
}
