% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ortho.R
\name{seq_ortho}
\alias{seq_ortho}
\title{Download RGB and/or IRC orthophotos for a Sequoia project}
\usage{
seq_ortho(
  dirname = ".",
  type = c("irc", "rgb"),
  buffer = 200,
  zoom = 12,
  crs = 2154,
  overwrite = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{dirname}{`character` Directory where the matrice file is located.
Defaults to the current working directory.}

\item{type}{`character` One or several orthophoto types to download.
Must be one or both of:
  - `"rgb"` - true-color orthophoto
  - `"irc"` - near-infrared orthophoto}

\item{buffer}{`numeric`; Buffer around `x` (in **meters**) used to enlarge
the download area.}

\item{zoom}{`integer` between 0 and 21. The smaller the zoom level, the less
precise the resolution(see [happign::get_wmts()])}

\item{crs}{`numeric` or `character`; CRS of the returned raster (see
[happign::get_wmts()])}

\item{overwrite}{`logical`; If `TRUE`, file is overwritten.}

\item{verbose}{`logical`; If `TRUE`, display messages.}
}
\value{
A named list of file paths written by [seq_write()], one per `type`.
}
\description{
Downloads one or several orthophotos (RGB and/or IRC) from the IGN WMTS service
for the `parca` layer of a Sequoia project.
}
\details{
This function is a convenience wrapper looping over [get_ortho()], allowing
the user to download both products in one call and automatically write them
to the project directory using [seq_write()].
}
\seealso{
[get_ortho()], [seq_write()]
}
