% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parca.R
\name{seq_parca}
\alias{seq_parca}
\title{Download, enrich and write cadastral geometries}
\usage{
seq_parca(dirname = ".", bdp_geom = TRUE, verbose = TRUE, overwrite = FALSE)
}
\arguments{
\item{dirname}{`character` Path to the directory. Defaults to the current
working directory.}

\item{bdp_geom}{`logical` If `TRUE`, replace Etalab geometries with BDP when possible.}

\item{verbose}{`logical` If `TRUE`, display messages.}

\item{overwrite}{`logical` If `TRUE`, filename is overwritten.}
}
\value{
An `sf` object
}
\description{
This function reads the `matrice.xlsx` file from a project directory,
downloads the geometry of each parcel, enriches the dataset with information
from the matrice, and writes the resulting layer to disk.
}
\details{
The resulting object is returned invisibly as an `sf` polygons layer.
The output file is automatically written into the working directory defined
by `dirname`.


**`bdp_geom`**
The use and behaviour of `bdp_geom` are described in [Rsequoia2::get_parca()].

**Automatic "lieu-dit" completion**
If the function detects rows in the matrice where the field `"LIEU_DIT"`
is missing, the corresponding "lieu-dit" values will be downloaded
automatically and added to the output.
Existing `"LIEU_DIT"` values are **never overwritten**.
}
