% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ua-parcels.R
\name{seq_parcels}
\alias{seq_parcels}
\title{Create _PF_ and _SSPF_ object from _UA_ for a Sequoia project}
\usage{
seq_parcels(dirname = ".", verbose = FALSE, overwrite = FALSE)
}
\arguments{
\item{dirname}{`character` Directory where the matrice file is located.
Defaults to the current working directory.}

\item{verbose}{`logical` If `TRUE`, display messages.}

\item{overwrite}{`logical` If `TRUE`, file is overwritten.}
}
\value{
A named list of six file paths, corresponding to the exported PF and SSPF
polygon and line layers.
}
\description{
This function reads the UA polygon layer (`v.seq.ua.poly`) of a Sequoia
project, aggregates surfaces at the `parcelle` and
`sous-parcelle` level and writes four derived layers:
}
\details{
Created layers are automatically written to the project directory using
[seq_write()], respecting the `dirname`, `overwrite`, and `verbose`
parameters.


The function performs two levels of spatial aggregation ande create 4 layers:

1. **PF (Parcelle)**
Rows of the UA are grouped using the `parcelle` field, and corrected surface
areas are summed. A polygon and border lines layer are created.

2. **SSPF (Sous-parcelle)**
Rows of the UA are grouped using the `parcelle` and `sous-parcelle` field, and
corrected surface areas are summed. A polygon and border lines
layer are created.
}
