% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parcels.R
\name{seq_parcels}
\alias{seq_parcels}
\title{Create \emph{PF} and \emph{SSPF} object from \emph{UA} for a Sequoia project}
\usage{
seq_parcels(dirname = ".", verbose = FALSE, overwrite = FALSE)
}
\arguments{
\item{verbose}{\code{logical} If \code{TRUE}, display messages.}

\item{overwrite}{\code{logical} If \code{TRUE}, file is overwritten.}
}
\value{
A named list of four file paths, corresponding to the exported PF and SSPF
polygon and line layers.
}
\description{
This function reads the UA polygon layer (\code{v.seq.ua.poly}) of a Sequoia
project, aggregates surfaces at the \code{parcelle} and
\code{sous-parcelle} level and writes four derived layers:
}
\details{
Created layers are automatically written to the project directory using
\code{\link[=seq_write]{seq_write()}}, respecting the \code{dirname}, \code{overwrite}, and \code{verbose}
parameters.

The function performs two levels of spatial aggregation ande create 4 layers:
\enumerate{
\item \strong{PF (Parcelle)}
Rows of the UA are grouped using the \code{parcelle} field, and corrected surface
areas are summed. A polygon and border lines layer are created.
\item \strong{SSPF (Sous-parcelle)}
Rows of the UA are grouped using the \code{parcelle} and \code{sous-parcelle} field, and
corrected surface areas are summed. A polygon and border lines
layer are created.
}
}
