% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patrimony.R
\name{seq_patrimony}
\alias{seq_patrimony}
\title{Download patrimony vector layers for a Sequoia project}
\usage{
seq_patrimony(
  dirname = ".",
  buffer = 500,
  key = get_keys("pat"),
  verbose = TRUE,
  overwrite = FALSE
)
}
\arguments{
\item{dirname}{`character` Directory where the matrice file is located.
Defaults to the current working directory.}

\item{buffer}{`numeric`; Buffer around `x` (in **meters**) used to enlarge
the download area.}

\item{key}{`character`; List of layer identifiers to download. If not
provided, the function uses `get_keys("pat")` to automatically select all
patrimony layers defined in the Sequoia configuration (`inst/config/seq_layers.yaml`)}

\item{verbose}{`logical`; If `TRUE`, display messages.}

\item{overwrite}{`logical` If `TRUE`, file is overwritten.}
}
\value{
A named list of file paths written by [seq_write()], one per layer.
}
\description{
Downloads one or several vector layers with `frheritage`
for `pat` layer(s) of a Sequoia project.
}
\details{
This function is a convenience wrapper looping over [get_patrimony()], allowing
the user to download all products in one call and automatically write them
to the project directory using [seq_write()].


For each value in `key`, the function attempts to query the corresponding
MNHN layer using [get_patrimony()].

- If the layer contains features, it is written to the project directory
  via [seq_write()] and recorded as a successful download.
- If the layer contains no features, it is skipped and marked
  as empty.
}
\seealso{
[get_patrimony()], [seq_write()]
}
