% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write.R
\name{seq_read}
\alias{seq_read}
\title{Open file based on key name}
\usage{
seq_read(key, dirname = ".", verbose = FALSE)
}
\arguments{
\item{key}{`character` Name of a layer key to match against the entries
defined in `inst/config/seq_layers.yaml`. (see *Details* for partial matching).}

\item{dirname}{`character` Directory where the matrice file is located.
Defaults to the current working directory.}

\item{verbose}{`logical` If `TRUE`, display messages.}
}
\value{
Object of class `SpatRaster` for raster and `sf` for vector
}
\description{
Open file based on key name
}
