% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write.R
\name{seq_read}
\alias{seq_read}
\title{Open file based on key name}
\usage{
seq_read(key, dirname = ".", verbose = FALSE)
}
\arguments{
\item{key}{\code{character} Name of a layer key to match against the entries
defined in \code{inst/config/seq_layer.yaml}. (see \emph{Details} for partial matching).}

\item{dirname}{\code{character} Directory where the matrice file is located.
Defaults to the current working directory.}

\item{verbose}{\code{logical} If \code{TRUE}, display messages.}
}
\value{
Object of class \code{SpatRaster} for raster and \code{sf} for vector
}
\description{
Open file based on key name
}
