% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{seq_retry}
\alias{seq_retry}
\title{Add retry capability to seq_* function}
\usage{
seq_retry(expr, times = 3, wait = 0.5, verbose = TRUE)
}
\arguments{
\item{expr}{Code to capture}

\item{times}{Number of retry}

\item{wait}{Time to wait between retry}

\item{verbose}{\code{logical} If \code{TRUE}, display messages.}
}
\description{
Internal helper used to retry download when failing
}
\keyword{internal}
