% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alti-rgealti.R
\name{seq_rgealti}
\alias{seq_rgealti}
\title{Create RGE ALTI layers for a Sequoia project}
\usage{
seq_rgealti(
  dirname = ".",
  key = c("mnt", "mns", "mnh"),
  buffer = 200,
  res = 1,
  crs = 2154,
  overwrite = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{dirname}{\code{character} Directory where the matrice file is located.
Defaults to the current working directory.}

\item{key}{\code{character}; RGE ALTI product(s) to create. One or more of
\code{"mnt"}, \code{"mns"} and \code{"mnh"}.}

\item{buffer}{\code{numeric}; Buffer around \code{x} (in \strong{meters}) used to enlarge
the download area.}

\item{res}{\code{numeric}; resolution specified in the units of the coordinate
system (see \code{\link[happign:get_wms_raster]{happign::get_wms_raster()}})}

\item{crs}{\code{numeric} or \code{character}; CRS of the returned raster (see
\code{\link[happign:get_wms_raster]{happign::get_wms_raster()}})}

\item{overwrite}{\code{logical} If \code{TRUE}, file is overwritten.}

\item{verbose}{\code{logical}; If \code{TRUE}, display messages.}
}
\value{
Invisibly returns a named \code{character} vector of output raster paths.
}
\description{
Uses the project's \emph{PARCA} layer to download and write RGE ALTI MNT, MNS
and/or MNH rasters.
}
