% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alti-rgealti.R
\name{seq_rgealti}
\alias{seq_rgealti}
\title{Create RGE ALTI layers for a Sequoia project}
\usage{
seq_rgealti(
  dirname = ".",
  key = c("mnt", "mns", "mnh"),
  buffer = 200,
  res = 1,
  crs = 2154,
  overwrite = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{dirname}{`character` Directory where the matrice file is located.
Defaults to the current working directory.}

\item{key}{`character`; RGE ALTI product(s) to create. One or more of
`"mnt"`, `"mns"` and `"mnh"`.}

\item{buffer}{`numeric`; Buffer around `x` (in **meters**) used to enlarge
the download area.}

\item{res}{`numeric`; resolution specified in the units of the coordinate
system (see [happign::get_wms_raster()])}

\item{crs}{`numeric` or `character`; CRS of the returned raster (see
[happign::get_wms_raster()])}

\item{overwrite}{`logical` If `TRUE`, file is overwritten.}

\item{verbose}{`logical`; If `TRUE`, display messages.}
}
\value{
Invisibly returns a named `character` vector of output raster paths.
}
\description{
Uses the project's _PARCA_ layer to download and write RGE ALTI MNT, MNS
and/or MNH rasters.
}
