% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{seq_summary}
\alias{seq_summary}
\title{Summarize information for a Sequoia folder}
\usage{
seq_summary(dirname = ".", verbose = TRUE, overwrite = FALSE)
}
\arguments{
\item{dirname}{\code{character} Directory where the matrice file is located.
Defaults to the current working directory.}

\item{verbose}{\code{logical} If \code{TRUE}, display messages.}

\item{overwrite}{\code{logical} If \code{TRUE}, file is overwritten.}
}
\value{
\code{list} of \code{data.frame}
}
\description{
Generate synthetic table from Sequoia layers and write them as \code{.xlsx}
inside the current Sequoia dir. Useful for redaction of management document.
}
