% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ua.R
\name{seq_ua}
\alias{seq_ua}
\title{Check and update the analysis units layer}
\usage{
seq_ua(
  dirname = ".",
  secure = TRUE,
  verbose = TRUE,
  overwrite = TRUE,
  check = interactive()
)
}
\arguments{
\item{dirname}{\code{character} Directory where the matrice file is located.
Defaults to the current working directory.}

\item{secure}{\code{logical}. If \code{TRUE}, also writes a timestamped secure copy
of the file.}

\item{verbose}{\code{logical} If \code{TRUE}, display messages.}

\item{overwrite}{\code{logical} If \code{TRUE}, file is overwritten.}

\item{check}{\code{logical} If \code{TRUE}, ask user.}
}
\value{
An \code{sf} object
}
\description{
This function reads the analysis units layer file from a project directory,
check and udapte attributes and overwrite it.
}
\details{
The resulting object is returned invisibly as an \code{sf} polygons layer.
The output file is automatically written into the working directory defined
by \code{dirname}.
}
