% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write.R
\name{seq_write}
\alias{seq_write}
\title{Write a spatial object based on a layer key}
\usage{
seq_write(x, key, dirname = ".", id = NULL, verbose = FALSE, overwrite = FALSE)
}
\arguments{
\item{x}{An `sf` object (for vector outputs) or a `SpatRaster` (for raster outputs).}

\item{key}{`character` Name of a layer key to match against the entries
defined in `inst/config/seq_layers.yaml`. (see *Details* for partial matching).}

\item{dirname}{`character` Directory where the matrice file is located.
Defaults to the current working directory.}

\item{id}{`character`. Project identifier used to prefix the output filename.
If `NULL`, the identifier is automatically inferred. The function first
attempts to extract a single, non-missing value from the `IDENTIFIANT`
field of layer `x`; if this fails, it then looks for a valid identifier in
a `*_matrice.xlsx` file found in `dirname`. If no unique identifier can be
determined, the filename is left unchanged.}

\item{verbose}{`logical` If `TRUE`, display messages.}

\item{overwrite}{`logical` If `TRUE`, file is overwritten.}
}
\value{
Invisibly returns the filepath used for writing.
}
\description{
Write a spatial object based on a layer key
}
