% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write.R
\name{seq_write}
\alias{seq_write}
\title{Write a spatial object based on a layer key}
\usage{
seq_write(x, key, dirname = ".", verbose = FALSE, overwrite = FALSE)
}
\arguments{
\item{x}{An \code{sf} object (for vector outputs) or a \code{SpatRaster} (for raster outputs).}

\item{key}{\code{character} Name of a layer key to match against the entries
defined in \code{inst/config/seq_layer.yaml}. (see \emph{Details} for partial matching).}

\item{dirname}{\code{character} Directory where the matrice file is located.
Defaults to the current working directory.}

\item{verbose}{\code{logical} If \code{TRUE}, display messages.}

\item{overwrite}{\code{logical} If \code{TRUE}, file is overwritten.}
}
\value{
Invisibly returns the filepath used for writing.
}
\description{
Write a spatial object based on a layer key
}
