% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-xlsx.R
\name{seq_xlsx}
\alias{seq_xlsx}
\title{Save Multiple Data Frames to an Excel Workbook with Styling}
\usage{
seq_xlsx(..., filename, data_table = FALSE, overwrite = FALSE, verbose = TRUE)
}
\arguments{
\item{...}{`data.frame` Each `data.frame` is wrote to a different sheet name.
If `...` contain named arg, name is used as sheet name else variable name is
used.}

\item{filename}{`character` File path where the workbook will be saved.}

\item{data_table}{`logical` If `TRUE`, data table is set up with total row.
Default to `FALSE`.}

\item{overwrite}{`logical` If `TRUE`, filename is overwritten.}

\item{verbose}{`logical` If `TRUE`, display messages.}
}
\value{
Invisibly returns the path `filename` after saving.
}
\description{
Creates a new Excel workbook, adds each element of a named list of data frames
as a separate sheet, writes the data, and applies `style_table()` to each sheet.
}
\examples{
\dontrun{
library(openxlsx2)
df1 <- data.frame(A = 1:3, B = 4:6)
df2 <- data.frame(X = letters[1:3], Y = runif(3))

Create xlsx with two sheet named THIS_DF1 and THIS_DF2
save_tables_to_xlsx(THIS_DF1 = df1, THIS_DF2 = df2, filename = "my_data.xlsx")

Create xlsx with two sheet named df1 and df2
save_tables_to_xlsx(df1, df2, filename = "my_data.xlsx")

Create xlsx with two sheet named df1 and THIS_DF2
save_tables_to_xlsx(df1, THIS_DF2 = df2, filename = "my_data.xlsx")

}
}
