% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequoia.R
\name{sequoia2}
\alias{sequoia2}
\title{Rsequoia2 Main Interactive Workflow}
\usage{
sequoia2(overwrite = FALSE)
}
\arguments{
\item{overwrite}{\code{logical} If \code{TRUE}, file is overwritten.}
}
\value{
Invisibly returns current Sequoia path.
The function primarily calls other Rsequoia2 functions.
}
\description{
Launches the main interactive workflow for Rsequoia2, allowing the user to
select a project folder and choose among map layer creation or cartographic tools.
}
