% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ua.R
\name{ua_check_coverage}
\alias{ua_check_coverage}
\title{Check spatial coverage between \emph{UA} and \emph{PARCA}}
\usage{
ua_check_coverage(ua, parca)
}
\arguments{
\item{ua}{\code{sf} Object containing analysis units.}

\item{parca}{\code{sf} Object containing cadastral parcels.}
}
\value{
The input \code{ua} object, invisibly, if spatial coverage is valid.
Aborts otherwise.
}
\description{
Checks that every \emph{UA} feature intersects at least one \emph{PARCA} feature, and
that every \emph{PARCA} feature intersects at least one \emph{UA} feature.
}
