% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ua.R
\name{ua_check_idu}
\alias{ua_check_idu}
\title{Check cadastral IDU consistency between _UA_ and _PARCA_}
\usage{
ua_check_idu(ua, parca, verbose = FALSE)
}
\arguments{
\item{ua}{`sf` Object containing analysis units.}

\item{parca}{`sf` Object containing cadastral parcels.}

\item{verbose}{`logical` If `TRUE`, display messages.}
}
\value{
`TRUE` if IDU values are consistent; `FALSE` otherwise,
with CLI warnings.
}
\description{
Checks that all _PARCA_ IDU values are present in _UA_, and that all _UA_
IDU values exist in _PARCA_.
}
