% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ua.R
\name{ua_check_idu}
\alias{ua_check_idu}
\title{Check cadastral IDU consistency between \emph{UA} and \emph{PARCA} sf objects}
\usage{
ua_check_idu(ua, parca, verbose = FALSE)
}
\arguments{
\item{ua}{\code{sf} Object from \code{\link[=parca_to_ua]{parca_to_ua()}} containing analysis
units}

\item{parca}{\code{sf} Object from \code{\link[=seq_parca]{seq_parca()}} containing cadastral
parcels.}

\item{verbose}{\code{logical} If \code{TRUE}, display progress messages.}
}
\value{
\code{TRUE} if all \code{parca} idu values are found in \code{ua};
\code{FALSE} otherwise (with CLI messages if \code{verbose = TRUE}).
}
\description{
Check cadastral IDU consistency between \emph{UA} and \emph{PARCA} sf objects
}
