% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ua.R
\name{ua_check_idu}
\alias{ua_check_idu}
\title{Check cadastral IDU consistency between UA and PARCA sf objects}
\usage{
ua_check_idu(ua, parca, verbose = TRUE)
}
\arguments{
\item{ua}{\code{sf} object containing analysis units; must contain the idu field.}

\item{parca}{\code{sf} Object containing cadastral parcels; must contain the idu field.}

\item{verbose}{\code{logical} If \code{TRUE}, display progress messages.}
}
\value{
\code{TRUE} if all \code{parca} idu values are found in \code{ua};
\code{FALSE} otherwise (with CLI messages if \code{verbose = TRUE}).
}
\description{
Check cadastral IDU consistency between UA and PARCA sf objects
}
