% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ua-parca.R
\name{ua_generate_area}
\alias{ua_generate_area}
\title{Calculate areas in the _UA_ sf object}
\usage{
ua_generate_area(ua, verbose = TRUE)
}
\arguments{
\item{ua}{`sf` object containing analysis units.}

\item{verbose}{`logical` If `TRUE`, display progress messages.}
}
\value{
`sf` object with updated areas
}
\description{
Computes the corrected cadastral areas for units of analysis according to
cartographic area.
}
