% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ua-ug.R
\name{ua_generate_ug}
\alias{ua_generate_ug}
\title{Create management unit field (UG) in the _UA_ sf object}
\usage{
ua_generate_ug(ua, separator = ".", verbose = TRUE)
}
\arguments{
\item{ua}{`sf` object containing analysis units;
must contain fields used by `ug_keys`.}

\item{separator}{`character`, default `"."`. Separator between keys.}

\item{verbose}{`logical` If `TRUE`, display progress messages.}
}
\value{
`sf` object UA with UG field filled.
}
\description{
Generates a standardized management unit identifier (UG) in the _UA_ object
based on configured parcel keys.
}
