% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ua.R
\name{ua_generate_ug}
\alias{ua_generate_ug}
\title{Create management unit field (UG) in the UA sf object}
\usage{
ua_generate_ug(
  ua,
  ug_keys = c("parcelle", "sous_parcelle"),
  separator = ".",
  verbose = TRUE
)
}
\arguments{
\item{ua}{\code{sf} object containing analysis units;
must contain fields used by \code{ug_keys}.}

\item{ug_keys}{\code{character} vector, default \code{c("parcelle", "sous_parcelle")}.
Keys used to build the UG identifier.}

\item{separator}{\code{character}, default \code{"."}. Separator between keys.}

\item{verbose}{\code{logical} If \code{TRUE}, display progress messages.}
}
\value{
\code{sf} object UA with UG field filled.
}
\description{
Generates a standardized management unit identifier (UG) in the UA object
based on configured parcel keys.
}
