% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ua-occupation.R
\name{ua_repair_dgd}
\alias{ua_repair_dgd}
\title{Repair DGD submitted status}
\usage{
ua_repair_dgd(ua, verbose = TRUE)
}
\arguments{
\item{ua}{\code{sf} object containing UA polygons.}

\item{verbose}{Logical. Should messages be printed?}
}
\value{
The repaired \code{ua} object.
}
\description{
Initializes DGD submitted status to TRUE when the field is empty.
}
