% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ua-occupation.R
\name{ua_repair_wooded}
\alias{ua_repair_wooded}
\title{Repair wooded status from DGD status}
\usage{
ua_repair_wooded(ua, verbose = TRUE)
}
\arguments{
\item{ua}{\code{sf} object containing UA polygons.}

\item{verbose}{Logical. Should messages be printed?}
}
\value{
The repaired \code{ua} object.
}
\description{
Forces wooded status to FALSE when a surface is not submitted to DGD.
}
