% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ua.R
\name{ua_to_parca}
\alias{ua_to_parca}
\title{Create \emph{PARCA} sf object from \emph{UA} sf object}
\usage{
ua_to_parca(ua)
}
\arguments{
\item{ua}{\code{sf} object containing analysis units}
}
\value{
An \code{sf} object containing cadastral parcels.
}
\description{
Create \emph{PARCA} sf object from \emph{UA} sf object
}
\details{
This function deviates from the traditional process.
It allows the \emph{PARCA} layer to be recreated from a \emph{UA} layer
in the case of a folder where the first layer is missing.
}
