% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parcels.R
\name{ua_to_pf}
\alias{ua_to_pf}
\title{Aggregate \emph{UA} surfaces at the \code{parcelle} field level}
\usage{
ua_to_pf(ua)
}
\arguments{
\item{ua}{\code{sf} object containing analysis units}
}
\value{
\code{sf} object where polygons are aggregated at the \emph{PF} level, with
corrected surfaces.
}
\description{
Aggregates corrected surface areas (\code{surf_cor}) from a \emph{UA} layer by
grouping rows according to the \code{parcelle} field, as defined by the
configuration returned by \code{\link[=seq_field]{seq_field()}}.
}
\details{
This function is typically used internally by \code{\link[=seq_parcels]{seq_parcels()}}, but may also
be called directly when a PF-level summary of a UA dataset is required.
}
\seealso{
\code{\link[=ua_to_sspf]{ua_to_sspf()}}, \code{\link[=seq_parcels]{seq_parcels()}}, \code{\link[=seq_field]{seq_field()}}, \code{\link[=seq_normalize]{seq_normalize()}}
}
