% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parcels.R
\name{ua_to_sspf}
\alias{ua_to_sspf}
\title{Aggregate \emph{UA} surfaces at the \code{sspf} field level}
\usage{
ua_to_sspf(ua)
}
\arguments{
\item{ua}{\code{sf} object containing analysis units}
}
\value{
An \code{sf} object where polygons and descriptive fields are aggregated
at the \emph{SSPF} level.
}
\description{
Aggregates corrected surface areas (\code{surf_cor}) from a \emph{UA} layer by
grouping rows according to the \code{sspf} field, as
defined by the configuration returned by \code{\link[=seq_field]{seq_field()}}.
}
\details{
In addition to summing surfaces, descriptive fields are preserved.

This function is used internally by \code{\link[=seq_parcels]{seq_parcels()}}, but may be called
directly when a SSPF-level summary of UA data is needed.
}
\seealso{
\code{\link[=ua_to_pf]{ua_to_pf()}}, \code{\link[=seq_parcels]{seq_parcels()}}, \code{\link[=seq_field]{seq_field()}}, \code{\link[=seq_desc_fields]{seq_desc_fields()}}, \code{\link[=seq_normalize]{seq_normalize()}}
}
