% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wooded.R
\name{ua_to_wooded}
\alias{ua_to_wooded}
\title{Aggregate \emph{UA} surfaces by wooded / non-wooded status}
\usage{
ua_to_wooded(ua)
}
\arguments{
\item{ua}{\code{sf} object containing analysis units.}
}
\value{
An \code{sf} object where polygons are aggregated by wooded status, with
corrected surfaces.
}
\description{
Aggregates corrected surface areas from a \emph{UA} layer by
grouping rows according to the wooded status field, as defined by the
configuration returned by \code{\link[=seq_field]{seq_field()}}.
}
\details{
This function is typically used internally by \code{\link[=seq_wooded]{seq_wooded()}}, but may also
be called directly when a summary of wooded versus non-wooded areas is
required.
}
\seealso{
\code{\link[=seq_wooded]{seq_wooded()}}, \code{\link[=seq_field]{seq_field()}}, \code{\link[=seq_normalize]{seq_normalize()}}
}
