% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ua.R
\name{ua_update_parca_fields}
\alias{ua_update_parca_fields}
\title{Update \emph{UA} fields from \emph{PARCA}}
\usage{
ua_update_parca_fields(ua, parca)
}
\arguments{
\item{ua}{\code{sf} Object containing analysis units.}

\item{parca}{\code{sf} Object containing cadastral parcels.}
}
\value{
A \code{sf} \emph{UA} object with cadastral fields updated from \emph{PARCA}.
}
\description{
Refreshes \emph{UA} cadastral fields from \emph{PARCA} using the IDU field as key.
}
\details{
This function assumes that \emph{UA} IDU values are already correct.
}
