% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update_parca}
\alias{update_parca}
\title{Update \code{R_SEQUOIA} \emph{PARCA} layer to \code{Rsequoia2}}
\usage{
update_parca(parca, id)
}
\arguments{
\item{parca}{\code{sf} object containing cadastrals parcels}

\item{id}{\code{character} Identifier of the project}
}
\value{
An \code{sf} object containing cadastral parcels, normalized for
\code{Rsequoia2} use.
}
\description{
Update \code{R_SEQUOIA} \emph{PARCA} layer to \code{Rsequoia2}
}
